`Logos`: Extract pure helper clusters from main.py into sibling modules
main.py grew to 6147 lines with several self-contained helper clusters
defined inline. Move the clearly separable, pure ones out without
changing any behavior:

* live_stream.py — in-flight streaming state: _LiveStreamRegistry,
  _StreamingLogAccumulator and _usage_tokens_from_payload
* logosnode_snapshot.py — pure shaping of logosnode runtime snapshots:
  the _LOGOSNODE_* env constants, datetime/window parsing and the
  scheduler-signal / live-sample builders
* middleware.py — APIPrefixStripperMiddleware
* dbutils/dbrequest.py — the six internal endpoint request models
  (RefreshPipelineRequest, Internal{Calibrate,Delete,Add,Sleep,Wake}
  LaneRequest), renamed to match the module's public naming

The moved names are re-imported into main.py, so the public entry point
logos.main:app and every existing attribute access keep working. Four
test files get tiny renames; AGENTS.md documents the new layout.
main.py goes from 6147 to 5082 lines.

Fixes #794
