+++ logos/logos-orchestrator/src/logos/main.py
@@ -1,7 +1,6 @@
 import asyncio
 import base64
 import binascii
-import codecs
 import datetime
 import hmac
 import json
@@ -13,16 +12,14 @@ import secrets
 import threading
 import time
 from contextlib import aclosing, asynccontextmanager, suppress
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Any, Dict, Optional, Set
 
 import grpc
-from fastapi import FastAPI, HTTPException, Request, WebSocket, WebSocketDisconnect
-from fastapi.encoders import jsonable_encoder
+from fastapi import FastAPI, HTTPException, Request
 from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse, Response, StreamingResponse
-from pydantic import BaseModel, Field
 from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from grpclocal import model_pb2_grpc
@@ -34,16 +31,7 @@ from logos.benchmarks.guidellm_runner import (
     BENCHMARK_PHASE_HEADER,
     BENCHMARK_PROVIDER_HEADER,
     BENCHMARK_TOKEN_HEADER,
-)
-from logos.benchmarks.guidellm_runner import DATASET as BENCHMARK_DATASET
-from logos.benchmarks.guidellm_runner import (
-    benchmark_affinity_headers,
     benchmark_affinity_token,
-    credential_transport_is_secure,
-    extract_serving_configuration,
-    internal_benchmark_target,
-    resolve_benchmark_target,
-    run_benchmark_job,
 )
 from logos.billing.finalize import finalize_billing_inputs
 from logos.capacity.calibration_orchestrator import CalibrationConfig, CalibrationOrchestrator
@@ -52,7 +40,7 @@ from logos.capacity.demand_tracker import DemandTracker
 from logos.classification.classification_balancer import Balancer
 from logos.classification.classification_manager import ClassificationManager
 from logos.context_budget import estimate_prompt_tokens, required_context_tokens
-from logos.dbutils.dbmanager import DBManager, derived_reported_context_length
+from logos.dbutils.dbmanager import DBManager
 from logos.dbutils.dbmodules import JobStatus
 from logos.dbutils.dbrequest import *
 from logos.dbutils.types import (
@@ -63,13 +51,21 @@ from logos.dbutils.types import (
 )
 from logos.errors import UpstreamStreamError, coerce_upstream_error, openai_error_response
 from logos.jobs.job_service import JobService, JobSubmission
-from logos.logosnode_registry import (
-    LogosNodeCommandError,
-    LogosNodeOfflineError,
-    LogosNodeRuntimeRegistry,
-    LogosNodeSessionConflictError,
+from logos.live_stream import _LiveStreamRegistry, _StreamingLogAccumulator, _usage_tokens_from_payload
+from logos.logosnode_registry import LogosNodeCommandError, LogosNodeOfflineError, LogosNodeRuntimeRegistry
+from logos.logosnode_snapshot import (
+    _build_live_local_provider_sample,
+    _is_today_or_all_utc,
+    _lane_served_context_window,
+    _logosnode_snapshot_is_connected,
+    _merge_provider_samples,
+    _profile_native_context_length,
+    _resolve_requested_model_name,
+    _runtime_modes_for_lanes,
+    _safe_float,
+    _sample_snapshot_id,
 )
-from logos.monitoring.prometheus_metrics import metrics_response as _prometheus_metrics_response
+from logos.middleware import APIPrefixStripperMiddleware
 from logos.pipeline.context_resolver import ContextResolver
 from logos.pipeline.correcting_scheduler import ClassificationCorrectingScheduler
 from logos.pipeline.executor import ExecutionResult, Executor, StreamingExecutionStatus
@@ -90,7 +86,6 @@ from logos.request_content import (
     set_payload_field,
 )
 from logos.responses import extract_model, extract_service_tier, get_client_ip, request_setup
-from logos.role_auth import require_logos_admin_key
 from logos.sdi.azure_deployment_sync import AzureDeploymentSyncService
 from logos.sdi.azure_facade import AzureSchedulingDataFacade
 from logos.sdi.cloud_model_sync import CloudModelSyncService
@@ -110,9 +105,12 @@ from logos.terminal_logging import (
     style_model,
     style_request_id,
 )
-from logos.timeouts import global_timeout_s
-
-_SERVER_START_TIME = int(time.time())
+from logos.timeouts import (
+    _LOGOSNODE_INFER_TIMEOUT_SECONDS,
+    _LOGOSNODE_PRETOKEN_RETRIES,
+    _LOGOSNODE_PRETOKEN_RETRY_BACKOFF_S,
+    _LOGOSNODE_STREAM_TIMEOUT_SECONDS,
+)
 
 logger = logging.getLogger("LogosLogger")
 _grpc_server = None
@@ -138,12 +136,6 @@ def _cancel_benchmark_job(job_id: int, reason: str) -> bool:
     if task is not None and not task.done():
         task.cancel(reason)
     return cancelled
-
-
-def _cancel_benchmarks_for_changed_session(provider_id: int, session_id: str | None) -> None:
-    for job_id, (job_provider_id, expected_session_id) in list(_benchmark_sessions_by_job.items()):
-        if job_provider_id == provider_id and expected_session_id != session_id:
-            _cancel_benchmark_job(job_id, "Provider restarted or disconnected")
 
 
 def _resolve_provider_name(provider_id: int) -> str:
@@ -217,40 +209,6 @@ _azure_deployment_sync: Optional[AzureDeploymentSyncService] = None
 _cloud_model_sync: Optional[CloudModelSyncService] = None
 
 
-def _env_int(name: str, default: int) -> int:
-    raw = os.getenv(name)
-    if raw is None or not str(raw).strip():
-        return default
-    try:
-        return max(1, int(raw))
-    except (TypeError, ValueError):
-        return default
-
-
-# max(1, ...): a fractional LOGOS_TIMEOUT_S (e.g. 0.5) must not floor to 0 and
-# cause immediate timeouts — clamp to at least 1 second.
-_LOGOSNODE_INFER_TIMEOUT_SECONDS = max(1, int(global_timeout_s(_env_int("LOGOSNODE_INFER_TIMEOUT_SECONDS", 120))))
-_LOGOSNODE_STREAM_TIMEOUT_SECONDS = max(
-    1,
-    int(
-        global_timeout_s(
-            _env_int(
-                "LOGOSNODE_STREAM_TIMEOUT_SECONDS",
-                _LOGOSNODE_INFER_TIMEOUT_SECONDS,
-            )
-        )
-    ),
-)
-_LOGOSNODE_STATS_STALE_AFTER_SECONDS = _env_int("LOGOSNODE_STATS_STALE_AFTER_SECONDS", 30)
-# Transparent retry for a logosnode stream that fails BEFORE the first token is
-# forwarded to the client (e.g. a just-woken level-1 lane whose vLLM engine was
-# not yet serveable — the worker now fails cleanly before stream_start). Safe to
-# re-dispatch because nothing has been sent downstream yet; bounded, with a small
-# backoff so the lane finishes waking. Never retries once a token has streamed.
-_LOGOSNODE_PRETOKEN_RETRIES = _env_int("LOGOSNODE_PRETOKEN_RETRIES", 3)
-_LOGOSNODE_PRETOKEN_RETRY_BACKOFF_S = float(os.getenv("LOGOSNODE_PRETOKEN_RETRY_BACKOFF_S", "1.0") or 1.0)
-
-
 def _record_azure_rate_limits(
     scheduling_stats: Optional[Dict[str, Any]],
     headers: Dict[str, str],
@@ -279,565 +237,6 @@ def _record_azure_rate_limits(
 
     if provider_metrics:
         _pipeline.record_provider_metrics(request_id, provider_metrics)
-
-
-def _parse_iso_datetime(raw: Any) -> Optional[datetime.datetime]:
-    if not isinstance(raw, str) or not raw.strip():
-        return None
-    try:
-        return datetime.datetime.fromisoformat(raw.replace("Z", "+00:00"))
-    except ValueError:
-        return None
-
-
-def _is_today_or_all_utc(day: str) -> bool:
-    normalized = str(day or "").strip().lower()
-    if normalized == "all":
-        return True
-    return normalized == datetime.datetime.now(datetime.timezone.utc).strftime("%Y-%m-%d")
-
-
-def _logosnode_snapshot_is_connected(snapshot: Optional[Dict[str, Any]]) -> bool:
-    if not isinstance(snapshot, dict):
-        return False
-    last_heartbeat = _parse_iso_datetime(snapshot.get("last_heartbeat"))
-    if last_heartbeat is None:
-        return False
-    now = datetime.datetime.now(datetime.timezone.utc)
-    return (now - last_heartbeat) <= datetime.timedelta(seconds=_LOGOSNODE_STATS_STALE_AFTER_SECONDS)
-
-
-def _normalize_loaded_models(lanes: list[dict[str, Any]]) -> list[dict[str, Any]]:
-    deduped: dict[str, dict[str, Any]] = {}
-    for lane in lanes:
-        if not isinstance(lane, dict):
-            continue
-        lane_model = str(lane.get("model") or "").strip()
-        loaded_models = lane.get("loaded_models") or []
-        if not isinstance(loaded_models, list):
-            loaded_models = []
-        for item in loaded_models:
-            if not isinstance(item, dict):
-                continue
-            name = str(item.get("name") or lane_model).strip()
-            if not name:
-                continue
-            size_vram = int(item.get("size_vram") or 0)
-            size_bytes = int(item.get("size") or 0)
-            current = deduped.get(name)
-            candidate = {
-                "name": name,
-                "size": size_bytes,
-                "size_vram": size_vram,
-            }
-            if current is None or candidate["size_vram"] > current["size_vram"]:
-                deduped[name] = candidate
-    return sorted(deduped.values(), key=lambda item: item["name"].lower())
-
-
-def _planner_model_alias(model_name: str) -> str:
-    """Return the planner/worker-safe alias used in lane ids and logs."""
-    return str(model_name or "").strip().replace("/", "_").replace(":", "_").replace(" ", "_")
-
-
-def _resolve_requested_model_name(
-    requested_name: str,
-    available_models: list[Dict[str, Any]],
-) -> Optional[str]:
-    """Resolve a user-supplied model id to a canonical DB model name.
-
-    ``available_models`` are the accessible model rows (each with a ``name``
-    and, optionally, an ``aliases`` list of stored alternative names).
-
-    Matches, in order of precedence:
-    1. the canonical name as stored in the DB,
-    2. a stored alias of the model (e.g. a logical tag like
-       ``local-most-powerful`` that can be re-pointed at another model),
-    3. the planner-safe alias form where ``/``, ``:``, and spaces are
-       rewritten as underscores (lets users copy model ids from lane names
-       or worker logs without breaking access-controlled model lookup).
-
-    All matching is case-insensitive. Every level must resolve to a single
-    unambiguous model to be accepted: the schema does not enforce
-    case-insensitive uniqueness of model names, so two models whose names
-    only differ in case make a canonical request ambiguous, and a stored
-    alias that matches several models does not fall through to the planner
-    aliases (a stored name is an explicit assignment and wins over the
-    derived form). Ambiguous requests resolve to ``None``.
-    """
-    requested = str(requested_name or "").strip()
-    if not requested:
-        return None
-    requested_lc = requested.lower()
-
-    canonical_matches: set[str] = set()
-    stored_alias_matches: set[str] = set()
-    planner_alias_matches: set[str] = set()
-    for entry in available_models:
-        canonical = str((entry or {}).get("name") or "").strip()
-        if not canonical:
-            continue
-        if canonical.lower() == requested_lc:
-            canonical_matches.add(canonical)
-            continue
-
-        sanitized = _planner_model_alias(canonical)
-        if requested_lc in {sanitized.lower(), f"planner-{sanitized.lower()}"}:
-            planner_alias_matches.add(canonical)
-        for alias in entry.get("aliases") or []:
-            if str(alias).strip().lower() == requested_lc:
-                stored_alias_matches.add(canonical)
-
-    if len(canonical_matches) == 1:
-        return next(iter(canonical_matches))
-    if canonical_matches:
-        # duplicate normalized model names — no way to tell which one was meant
-        return None
-    if len(stored_alias_matches) == 1:
-        return next(iter(stored_alias_matches))
-    if stored_alias_matches:
-        # an ambiguous stored alias must not fall through to planner aliases
-        return None
-    if len(planner_alias_matches) == 1:
-        return next(iter(planner_alias_matches))
-    return None
-
-
-def _runtime_modes_for_lanes(lanes: list[dict[str, Any]]) -> list[str]:
-    modes: set[str] = set()
-    for lane in lanes:
-        if not isinstance(lane, dict):
-            continue
-        modes.add("vllm" if bool(lane.get("vllm")) else "ollama")
-    return sorted(modes)
-
-
-def _safe_float(value: Any) -> Optional[float]:
-    if value is None:
-        return None
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return None
-
-
-def _safe_int(value: Any) -> Optional[int]:
-    if value is None:
-        return None
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return None
-
-
-def _merge_histogram_buckets(target: dict[str, float], source: Any) -> None:
-    if not isinstance(source, dict):
-        return
-    for raw_bucket, raw_value in source.items():
-        bucket = str(raw_bucket).strip()
-        if not bucket:
-            continue
-        count = _safe_float(raw_value)
-        if count is None or count < 0:
-            continue
-        target[bucket] = target.get(bucket, 0.0) + count
-
-
-def _histogram_quantile_seconds(histogram: Any, quantile: float = 0.95) -> Optional[float]:
-    if not isinstance(histogram, dict) or not histogram:
-        return None
-
-    buckets: list[tuple[float, float]] = []
-    for raw_bucket, raw_count in histogram.items():
-        count = _safe_float(raw_count)
-        if count is None or count < 0:
-            continue
-        bucket_label = str(raw_bucket).strip()
-        if not bucket_label:
-            continue
-        if bucket_label == "+Inf":
-            upper_bound = float("inf")
-        else:
-            upper_bound = _safe_float(bucket_label)
-            if upper_bound is None:
-                continue
-        buckets.append((upper_bound, count))
-
-    if not buckets:
-        return None
-
-    buckets.sort(key=lambda item: item[0])
-    total_count = max(count for _upper, count in buckets)
-    if total_count <= 0:
-        return None
-
-    target = total_count * max(0.0, min(1.0, quantile))
-    previous_upper = 0.0
-    previous_count = 0.0
-
-    for upper_bound, cumulative_count in buckets:
-        if cumulative_count < target:
-            previous_upper = 0.0 if upper_bound == float("inf") else upper_bound
-            previous_count = cumulative_count
-            continue
-
-        if upper_bound == float("inf"):
-            return previous_upper if previous_upper > 0 else None
-
-        bucket_count = cumulative_count - previous_count
-        if bucket_count <= 0:
-            return upper_bound
-
-        bucket_width = upper_bound - previous_upper
-        if bucket_width <= 0:
-            return upper_bound
-
-        fraction = (target - previous_count) / bucket_count
-        return previous_upper + (fraction * bucket_width)
-
-    last_upper = buckets[-1][0]
-    if last_upper == float("inf"):
-        return previous_upper if previous_upper > 0 else None
-    return last_upper
-
-
-def _build_logosnode_scheduler_signals(runtime: Dict[str, Any]) -> Dict[str, Any]:
-    if not isinstance(runtime, dict):
-        return {}
-
-    devices = runtime.get("devices") if isinstance(runtime.get("devices"), dict) else {}
-    capacity = runtime.get("capacity") if isinstance(runtime.get("capacity"), dict) else {}
-    transport = runtime.get("transport") if isinstance(runtime.get("transport"), dict) else {}
-    lanes = runtime.get("lanes") if isinstance(runtime.get("lanes"), list) else []
-    # Needed to resolve a lane's served window: a vLLM lane that was started
-    # without an explicit --max-model-len takes it from the calibrated profile,
-    # so the number is not on the lane itself.
-    model_profiles = runtime.get("model_profiles") if isinstance(runtime.get("model_profiles"), dict) else {}
-
-    provider_signals: Dict[str, Any] = {
-        "timestamp": runtime.get("timestamp"),
-        "transport_connected": bool(transport.get("connected", True)),
-        "device_mode": devices.get("mode"),
-        "nvidia_smi_available": bool(devices.get("nvidia_smi_available", False)),
-        "telemetry_available": bool(devices.get("telemetry_available", devices.get("nvidia_smi_available", False))),
-        "device_count": (len(devices.get("devices") or []) if isinstance(devices.get("devices"), list) else 0),
-        "total_memory_mb": _safe_float(devices.get("total_memory_mb")),
-        "used_memory_mb": _safe_float(devices.get("used_memory_mb")),
-        "free_memory_mb": _safe_float(devices.get("free_memory_mb")),
-        "lane_count": _safe_int(capacity.get("lane_count")) or len(lanes),
-        "active_requests": _safe_int(capacity.get("active_requests")) or 0,
-        "loaded_lane_count": _safe_int(capacity.get("loaded_lane_count")) or 0,
-        "sleeping_lane_count": _safe_int(capacity.get("sleeping_lane_count")) or 0,
-        "cold_lane_count": _safe_int(capacity.get("cold_lane_count")) or 0,
-        "total_effective_vram_mb": _safe_float(capacity.get("total_effective_vram_mb")) or 0.0,
-        "runtime_modes": _runtime_modes_for_lanes(lanes),
-    }
-
-    raw_device_list = devices.get("devices") or []
-    provider_signals["devices"] = [
-        {
-            "device_id": d.get("device_id", ""),
-            "kind": d.get("kind", "nvidia"),
-            "name": d.get("name", ""),
-            "memory_used_mb": float(d.get("memory_used_mb") or 0.0),
-            "memory_total_mb": float(d.get("memory_total_mb") or 0.0),
-            "memory_free_mb": float(d.get("memory_free_mb") or 0.0),
-            "utilization_percent": _safe_float(d.get("utilization_percent")),
-            "temperature_celsius": _safe_float(d.get("temperature_celsius")),
-            "power_draw_watts": _safe_float(d.get("power_draw_watts")),
-        }
-        for d in raw_device_list
-        if isinstance(d, dict)
-    ]
-
-    model_signals: dict[str, Dict[str, Any]] = {}
-    lane_signals: dict[str, Dict[str, Any]] = {}
-
-    def _ensure_model_entry(model_name: str) -> Dict[str, Any]:
-        return model_signals.setdefault(
-            model_name,
-            {
-                "lane_count": 0,
-                "vllm_lane_count": 0,
-                "ollama_lane_count": 0,
-                "loaded_lane_count": 0,
-                "running_lane_count": 0,
-                "sleeping_lane_count": 0,
-                "cold_lane_count": 0,
-                "starting_lane_count": 0,
-                "error_lane_count": 0,
-                "active_requests": 0,
-                "effective_vram_mb": 0.0,
-                "reported_vram_mb": 0.0,
-                "pid_vram_mb": 0.0,
-                "device_vram_mb": 0.0,
-                "queue_waiting_current": 0.0,
-                "requests_running_current": 0.0,
-                "prompt_tokens_total": None,
-                "generation_tokens_total": None,
-                "ttft_histogram": {},
-                "ttft_p95_seconds": None,
-                "gpu_cache_usage_percent_avg": None,
-                "gpu_cache_usage_percent_max": None,
-                "prefix_cache_hit_rate_avg": None,
-                "mtp_acceptance_rate_avg": None,
-                "_gpu_cache_usage_percent_sum": 0.0,
-                "_gpu_cache_usage_percent_count": 0,
-                "_prefix_cache_hit_rate_sum": 0.0,
-                "_prefix_cache_hit_rate_count": 0,
-                "_mtp_draft_tokens_total": 0.0,
-                "_mtp_accepted_tokens_total": 0.0,
-            },
-        )
-
-    for lane in lanes:
-        if not isinstance(lane, dict):
-            continue
-
-        model_name = str(lane.get("model") or "").strip()
-        lane_id = str(lane.get("lane_id") or "").strip()
-        runtime_state = str(lane.get("runtime_state") or "").strip()
-        is_vllm = bool(lane.get("vllm"))
-        active_requests = _safe_int(lane.get("active_requests")) or 0
-        backend_metrics = lane.get("backend_metrics") if isinstance(lane.get("backend_metrics"), dict) else {}
-        ttft_histogram = (
-            backend_metrics.get("ttft_histogram") if isinstance(backend_metrics.get("ttft_histogram"), dict) else {}
-        )
-        lane_ttft_p95 = _histogram_quantile_seconds(ttft_histogram)
-
-        lane_signal = {
-            "model": model_name,
-            "vllm": is_vllm,
-            "runtime_state": runtime_state,
-            "sleep_state": lane.get("sleep_state"),
-            "gpu_devices": str(lane.get("gpu_devices") or ""),
-            "effective_gpu_devices": str(lane.get("effective_gpu_devices") or ""),
-            "num_parallel": _safe_int(lane.get("num_parallel")) or 0,
-            "active_requests": active_requests,
-            "effective_vram_mb": _safe_float(lane.get("effective_vram_mb")) or 0.0,
-            "reported_vram_mb": _safe_float(lane.get("reported_vram_mb")) or 0.0,
-            "pid_vram_mb": _safe_float(lane.get("pid_vram_mb")) or 0.0,
-            "device_vram_mb": _safe_float(lane.get("device_vram_mb")) or 0.0,
-            "vram_source": lane.get("vram_source"),
-            "queue_waiting": _safe_float(backend_metrics.get("queue_waiting")),
-            "requests_running": _safe_float(backend_metrics.get("requests_running")),
-            "gpu_cache_usage_percent": _safe_float(backend_metrics.get("gpu_cache_usage_percent")),
-            "prefix_cache_hit_rate": _safe_float(backend_metrics.get("prefix_cache_hit_rate")),
-            "mtp_acceptance_rate": _safe_float(backend_metrics.get("mtp_acceptance_rate")),
-            "prompt_tokens_total": _safe_float(backend_metrics.get("prompt_tokens_total")),
-            "generation_tokens_total": _safe_float(backend_metrics.get("generation_tokens_total")),
-            "ttft_histogram": ttft_histogram,
-            "ttft_p95_seconds": lane_ttft_p95,
-            # The window this lane is actually serving at. Two lanes of the same
-            # model can differ — the planner sizes each against the KV cache it
-            # could get — so it belongs on the lane and not on the model. 0 when
-            # the worker reports nothing to derive it from.
-            "max_model_len": _lane_served_context_window(lane, model_profiles) or None,
-        }
-        if lane_id:
-            lane_signals[lane_id] = lane_signal
-
-        if not model_name:
-            continue
-
-        entry = _ensure_model_entry(model_name)
-        entry["lane_count"] += 1
-        if is_vllm:
-            entry["vllm_lane_count"] += 1
-        else:
-            entry["ollama_lane_count"] += 1
-
-        if runtime_state == "loaded":
-            entry["loaded_lane_count"] += 1
-        elif runtime_state == "running":
-            entry["running_lane_count"] += 1
-        elif runtime_state == "sleeping":
-            entry["sleeping_lane_count"] += 1
-        elif runtime_state == "cold":
-            entry["cold_lane_count"] += 1
-        elif runtime_state == "starting":
-            entry["starting_lane_count"] += 1
-        elif runtime_state == "error":
-            entry["error_lane_count"] += 1
-
-        entry["active_requests"] += active_requests
-        entry["effective_vram_mb"] += _safe_float(lane.get("effective_vram_mb")) or 0.0
-        entry["reported_vram_mb"] += _safe_float(lane.get("reported_vram_mb")) or 0.0
-        entry["pid_vram_mb"] += _safe_float(lane.get("pid_vram_mb")) or 0.0
-        entry["device_vram_mb"] += _safe_float(lane.get("device_vram_mb")) or 0.0
-
-        queue_waiting = _safe_float(backend_metrics.get("queue_waiting"))
-        if queue_waiting is not None:
-            entry["queue_waiting_current"] += queue_waiting
-
-        requests_running = _safe_float(backend_metrics.get("requests_running"))
-        if requests_running is not None:
-            entry["requests_running_current"] += requests_running
-
-        prompt_tokens_total = _safe_float(backend_metrics.get("prompt_tokens_total"))
-        if prompt_tokens_total is not None:
-            current_prompt = _safe_float(entry.get("prompt_tokens_total")) or 0.0
-            entry["prompt_tokens_total"] = current_prompt + prompt_tokens_total
-
-        generation_tokens_total = _safe_float(backend_metrics.get("generation_tokens_total"))
-        if generation_tokens_total is not None:
-            current_generation = _safe_float(entry.get("generation_tokens_total")) or 0.0
-            entry["generation_tokens_total"] = current_generation + generation_tokens_total
-
-        gpu_cache_usage_percent = _safe_float(backend_metrics.get("gpu_cache_usage_percent"))
-        if gpu_cache_usage_percent is not None:
-            entry["_gpu_cache_usage_percent_sum"] += gpu_cache_usage_percent
-            entry["_gpu_cache_usage_percent_count"] += 1
-            current_max = _safe_float(entry.get("gpu_cache_usage_percent_max"))
-            entry["gpu_cache_usage_percent_max"] = (
-                gpu_cache_usage_percent if current_max is None else max(current_max, gpu_cache_usage_percent)
-            )
-
-        prefix_cache_hit_rate = _safe_float(backend_metrics.get("prefix_cache_hit_rate"))
-        if prefix_cache_hit_rate is not None:
-            entry["_prefix_cache_hit_rate_sum"] += prefix_cache_hit_rate
-            entry["_prefix_cache_hit_rate_count"] += 1
-
-        # Token-weighted MTP aggregation: sum the cumulative draft/accepted
-        # token counters per model (an unweighted mean of per-lane rates would
-        # misstate the model rate when lanes have different draft volumes).
-        mtp_draft_tokens_total = _safe_float(backend_metrics.get("mtp_draft_tokens_total"))
-        if mtp_draft_tokens_total is not None:
-            entry["_mtp_draft_tokens_total"] += mtp_draft_tokens_total
-
-        mtp_accepted_tokens_total = _safe_float(backend_metrics.get("mtp_accepted_tokens_total"))
-        if mtp_accepted_tokens_total is not None:
-            entry["_mtp_accepted_tokens_total"] += mtp_accepted_tokens_total
-
-        _merge_histogram_buckets(entry["ttft_histogram"], ttft_histogram)
-
-    for entry in model_signals.values():
-        gpu_count = int(entry.pop("_gpu_cache_usage_percent_count", 0) or 0)
-        gpu_sum = float(entry.pop("_gpu_cache_usage_percent_sum", 0.0) or 0.0)
-        if gpu_count > 0:
-            entry["gpu_cache_usage_percent_avg"] = gpu_sum / gpu_count
-
-        prefix_count = int(entry.pop("_prefix_cache_hit_rate_count", 0) or 0)
-        prefix_sum = float(entry.pop("_prefix_cache_hit_rate_sum", 0.0) or 0.0)
-        if prefix_count > 0:
-            entry["prefix_cache_hit_rate_avg"] = prefix_sum / prefix_count
-
-        mtp_draft = float(entry.pop("_mtp_draft_tokens_total", 0.0) or 0.0)
-        mtp_accepted = float(entry.pop("_mtp_accepted_tokens_total", 0.0) or 0.0)
-        if mtp_draft > 0:
-            entry["mtp_acceptance_rate_avg"] = mtp_accepted / mtp_draft
-
-        entry["ttft_p95_seconds"] = _histogram_quantile_seconds(entry.get("ttft_histogram"))
-
-    return {
-        "provider": provider_signals,
-        "models": model_signals,
-        "lanes": lane_signals,
-    }
-
-
-def _build_live_local_provider_sample(
-    provider: Optional[Dict[str, Any]],
-    snapshot: Optional[Dict[str, Any]],
-) -> Optional[Dict[str, Any]]:
-    if not _logosnode_snapshot_is_connected(snapshot):
-        return None
-
-    runtime = snapshot.get("runtime") if isinstance(snapshot, dict) else {}
-    if not isinstance(runtime, dict):
-        return None
-
-    lanes = runtime.get("lanes") if isinstance(runtime.get("lanes"), list) else []
-    devices = runtime.get("devices") if isinstance(runtime.get("devices"), dict) else {}
-    capacity = runtime.get("capacity") if isinstance(runtime.get("capacity"), dict) else {}
-    transport = runtime.get("transport") if isinstance(runtime.get("transport"), dict) else {}
-
-    used_vram_mb = float(devices.get("used_memory_mb") or 0.0)
-    if used_vram_mb <= 0:
-        used_vram_mb = float(capacity.get("total_effective_vram_mb") or 0.0)
-
-    total_vram_mb = float(devices.get("total_memory_mb") or 0.0)
-    if total_vram_mb <= 0 and isinstance(provider, dict) and provider.get("total_vram_mb") is not None:
-        total_vram_mb = float(provider.get("total_vram_mb") or 0.0)
-
-    remaining_vram_mb: Optional[float] = None
-    # telemetry_available is the backend-neutral successor to
-    # nvidia_smi_available: it means "the worker measured this on real
-    # hardware", whether that hardware reports through nvidia-smi or through
-    # Metal's device_info. Metal workers set it and leave nvidia_smi_available
-    # False, so gating on the old field alone would discard their free-memory
-    # reading and fall back to the coarser total-minus-used estimate.
-    if devices.get("telemetry_available") or devices.get("nvidia_smi_available"):
-        remaining_vram_mb = float(devices.get("free_memory_mb") or 0.0)
-    elif total_vram_mb > 0:
-        remaining_vram_mb = max(total_vram_mb - used_vram_mb, 0.0)
-
-    loaded_models = _normalize_loaded_models(lanes)
-    runtime_modes = _runtime_modes_for_lanes(lanes)
-    scheduler_signals = _build_logosnode_scheduler_signals(runtime)
-
-    if remaining_vram_mb is None and not loaded_models and used_vram_mb <= 0:
-        return None
-
-    # Pick the freshest timestamp the worker has given us. With an idle
-    # cluster the cached `runtime.timestamp` can be hours old (the worker
-    # only emits status on lane-state changes), while `last_heartbeat`
-    # advances on every WS heartbeat. Using the older of the two would
-    # collide with the already-persisted DB sample inside
-    # _merge_provider_samples and add no new point — leaving the chart
-    # with a single dot that scatter `mode: "lines"` can't draw.
-    runtime_ts = runtime.get("timestamp") if isinstance(runtime.get("timestamp"), str) else None
-    heartbeat_ts = snapshot.get("last_heartbeat") if isinstance(snapshot, dict) else None
-    if isinstance(heartbeat_ts, datetime.datetime):
-        heartbeat_ts = heartbeat_ts.isoformat()
-    elif not isinstance(heartbeat_ts, str):
-        heartbeat_ts = None
-    candidates = [t for t in (heartbeat_ts, runtime_ts) if isinstance(t, str) and t.strip()]
-    timestamp = max(candidates) if candidates else datetime.datetime.now(datetime.timezone.utc).isoformat()
-
-    return {
-        "timestamp": timestamp,
-        "snapshot_source": "logosnode-runtime",
-        "provider_type": (provider.get("provider_type") if isinstance(provider, dict) else "logosnode"),
-        "connection_state": "online",
-        "connected": True,
-        "transport_connected": bool(transport.get("connected", True)),
-        "runtime_modes": runtime_modes,
-        "vram_mb": used_vram_mb,
-        "used_vram_mb": used_vram_mb,
-        "remaining_vram_mb": remaining_vram_mb,
-        "total_vram_mb": total_vram_mb if total_vram_mb > 0 else None,
-        "models_loaded": len(loaded_models),
-        "loaded_models": loaded_models,
-        "runtime_payload": runtime,
-        "scheduler_signals": scheduler_signals,
-    }
-
-
-def _sample_snapshot_id(sample: Dict[str, Any]) -> int:
-    try:
-        return int(sample.get("snapshot_id") or 0)
-    except (TypeError, ValueError):
-        return 0
-
-
-def _sample_sort_key(sample: Dict[str, Any]) -> tuple[int, str]:
-    return (_sample_snapshot_id(sample), str(sample.get("timestamp") or ""))
-
-
-def _merge_provider_samples(
-    existing_samples: list[Dict[str, Any]],
-    extra_samples: list[Dict[str, Any]],
-) -> list[Dict[str, Any]]:
-    by_key: dict[str, Dict[str, Any]] = {}
-    for sample in list(existing_samples) + list(extra_samples):
-        if not isinstance(sample, dict):
-            continue
-        key = str(sample.get("snapshot_id") or sample.get("timestamp") or "")
-        if not key:
-            continue
-        by_key[key] = {**by_key.get(key, {}), **sample}
-    return sorted(by_key.values(), key=_sample_sort_key)
 
 
 def _load_persisted_local_provider_vram_payload(
@@ -875,96 +274,6 @@ def _load_persisted_local_provider_vram_payload(
     payload.setdefault("providers", [])
     payload.setdefault("last_snapshot_id", int(after_snapshot_id or 0))
     return payload
-
-
-def _capture_logosnode_provider_snapshot(
-    provider_id: int,
-    runtime: Dict[str, Any],
-) -> None:
-    sample = _build_live_local_provider_sample(
-        None,
-        {
-            "last_heartbeat": runtime.get("timestamp"),
-            "runtime": runtime,
-        },
-    )
-    if sample is None:
-        return
-
-    timestamp = _parse_iso_datetime(sample.get("timestamp"))
-    used_bytes = int(float(sample.get("used_vram_mb") or 0.0) * 1024 * 1024)
-    total_vram_mb = sample.get("total_vram_mb")
-    total_bytes = None
-    if total_vram_mb is not None:
-        total_bytes = int(float(total_vram_mb or 0.0) * 1024 * 1024)
-    free_vram_mb = sample.get("remaining_vram_mb")
-    free_bytes = None
-    if free_vram_mb is not None:
-        free_bytes = int(float(free_vram_mb or 0.0) * 1024 * 1024)
-
-    with DBManager() as db:
-        snapshot_id = db.insert_provider_snapshot(
-            provider_id=provider_id,
-            snapshot_ts=timestamp,
-            total_models_loaded=int(sample.get("models_loaded") or 0),
-            total_vram_used_bytes=used_bytes,
-            total_memory_bytes=total_bytes,
-            free_memory_bytes=free_bytes,
-            loaded_models=list(sample.get("loaded_models") or []),
-            snapshot_source=str(sample.get("snapshot_source") or "logosnode-runtime"),
-            runtime_payload=(sample.get("runtime_payload") if isinstance(sample.get("runtime_payload"), dict) else {}),
-            scheduler_signals=(
-                sample.get("scheduler_signals") if isinstance(sample.get("scheduler_signals"), dict) else {}
-            ),
-            poll_success=True,
-        )
-        # Persist calibrated model profiles into the dedicated table
-        runtime_payload = sample.get("runtime_payload")
-        if isinstance(runtime_payload, dict):
-            model_profiles = runtime_payload.get("model_profiles")
-            if isinstance(model_profiles, dict) and model_profiles:
-                try:
-                    db.upsert_model_profiles(provider_id, model_profiles)
-                except Exception:
-                    db.session.rollback()
-                    logger.warning(
-                        "Failed to upsert model profiles for provider %s, the "
-                        "entire row update (base_residency_mb, loaded_vram_mb, "
-                        "kv_budget_mb, measurement_count, last_measured_at) is "
-                        "lost until this recovers",
-                        _resolve_provider_name(provider_id),
-                        exc_info=True,
-                    )
-
-    sample["snapshot_id"] = snapshot_id
-    asyncio.create_task(_logosnode_registry.record_runtime_sample(provider_id, sample))
-
-
-def _capture_calibration_probe_log(provider_id: int, event: Dict[str, Any]) -> None:
-    """Persist a worker's ``calibration_probe_log`` event into the DB.
-
-    Fired once per model per calibration attempt (see
-    ``LogosBridgeClient._record_calibration_probe_log`` on the worker side).
-    Keeps only the most recent row per (provider_id, model_name) via
-    ``upsert_calibration_probe_log``'s ON CONFLICT — mirrors how
-    ``upsert_model_profiles`` above keeps one row per (provider_id,
-    model_name).
-    """
-    model_name = str(event.get("model") or "").strip()
-    if not model_name:
-        return
-    recorded_at = _parse_iso_datetime(event.get("timestamp"))
-    details_raw = event.get("details")
-    payload = json.loads(details_raw) if isinstance(details_raw, str) and details_raw else {}
-    if not isinstance(payload, dict):
-        return
-    # Pop out before it goes into `summary` below — otherwise the (large,
-    # deliberately un-truncated) raw log text would be duplicated into both
-    # the dedicated `log_text` column and the JSONB summary blob.
-    log_text = payload.pop("log_text", None)
-
-    with DBManager() as db:
-        db.upsert_calibration_probe_log(provider_id, model_name, recorded_at, payload, log_text)
 
 
 def _merge_local_provider_vram_payload(
@@ -1188,429 +497,7 @@ def _record_log_failure(
         )
 
 
-# Anthropic Messages SSE event types the accumulator acts on. The stream also
-# emits content_block_start/stop and ping, which carry neither text nor usage.
-_MESSAGES_EVENT_TYPES = frozenset({"message_start", "message_delta", "message_stop", "content_block_delta"})
-
-
-class _LiveStreamRegistry:
-    """Token counts of the requests running right now, keyed by request id.
-
-    A finished request's usage is in the database; one still running is
-    nowhere, and "nowhere" is what the statistics page showed for the whole
-    minute a long generation takes. This holds the in-flight view, in memory
-    and on the orchestrator, because that is the only process that sees the
-    chunks go past.
-
-    Bounded by construction: an entry exists only while its request is in
-    flight and is dropped in the same ``finally`` that logs the completion.
-    """
-
-    def __init__(self, now: Any = time.time) -> None:
-        self._streams: dict[str, dict[str, Any]] = {}
-        self._lock = threading.Lock()
-        # Bumped on every mutation so a subscriber (the SSE stream) can tell
-        # that something moved without diffing the whole table.
-        self._version = 0
-        # Injected rather than read off the module so a test can drive the clock
-        # without patching time.time for everything else running in-process.
-        self._now = now
-
-    @property
-    def version(self) -> int:
-        with self._lock:
-            return self._version
-
-    def start(
-        self,
-        request_id: Optional[str],
-        model_name: Optional[str] = None,
-        prompt_tokens: int = 0,
-        prompt_estimated: bool = False,
-    ) -> None:
-        """Register a request as in-flight; non-destructive for known ids.
-
-        A request is tracked from the moment it arrives, long before its
-        response starts streaming: while it queues, the only figure available
-        is the prompt estimate the context routing already computes. When the
-        streamer later calls this with the model name, the entry must keep its
-        counters and start time rather than resetting them — the numbers on
-        the page would jump backwards.
-        """
-        if not request_id:
-            return
-        with self._lock:
-            entry = self._streams.get(request_id)
-            if entry is None:
-                self._streams[request_id] = {
-                    "model_name": model_name,
-                    "started_at": self._now(),
-                    # Set on the first delta, not here: the wait for a lane to warm
-                    # up is not generation, and averaging over it would report a
-                    # rate no one is seeing.
-                    "first_token_at": None,
-                    "prompt_tokens": prompt_tokens,
-                    "completion_tokens": 0,
-                    # True until the upstream states the real prompt size; the
-                    # page shows such a figure as an estimate, not a fact.
-                    "prompt_estimated": prompt_estimated and prompt_tokens > 0,
-                }
-            else:
-                if model_name:
-                    entry["model_name"] = model_name
-                if prompt_tokens > 0 and entry["prompt_tokens"] == 0:
-                    entry["prompt_tokens"] = prompt_tokens
-                    entry["prompt_estimated"] = prompt_estimated
-            self._version += 1
-
-    def update(self, request_id: Optional[str], tokens: Dict[str, int]) -> None:
-        if not request_id:
-            return
-        with self._lock:
-            entry = self._streams.get(request_id)
-            if entry is None:
-                return
-            completion = tokens.get("completion_tokens", 0)
-            if completion > 0 and entry["first_token_at"] is None:
-                entry["first_token_at"] = self._now()
-            prompt = tokens.get("prompt_tokens", 0)
-            if prompt > 0:
-                # The upstream has stated the prompt size; the estimate it
-                # stood in for is no longer an estimate.
-                entry["prompt_tokens"] = prompt
-                entry["prompt_estimated"] = False
-            entry["completion_tokens"] = completion
-            self._version += 1
-
-    def finish(self, request_id: Optional[str]) -> None:
-        if not request_id:
-            return
-        with self._lock:
-            if self._streams.pop(request_id, None) is not None:
-                self._version += 1
-
-    def snapshot(self) -> list[dict[str, Any]]:
-        """Every running request, with the rate it is generating at."""
-        now = self._now()
-        out: list[dict[str, Any]] = []
-        with self._lock:
-            entries = [(rid, dict(entry)) for rid, entry in self._streams.items()]
-        for request_id, entry in entries:
-            first_token_at = entry["first_token_at"]
-            completion = entry["completion_tokens"]
-            elapsed = (now - first_token_at) if first_token_at else 0.0
-            out.append(
-                {
-                    "request_id": request_id,
-                    "model_name": entry["model_name"],
-                    "prompt_tokens": entry["prompt_tokens"],
-                    "prompt_estimated": entry["prompt_estimated"],
-                    "completion_tokens": completion,
-                    # Measured from the first token, so it is the generation
-                    # rate rather than an average dragged down by queueing.
-                    # None until there is a span to divide by.
-                    "tokens_per_second": (completion / elapsed) if elapsed > 0.5 and completion > 0 else None,
-                    "elapsed_seconds": now - entry["started_at"],
-                }
-            )
-        return out
-
-
 _live_streams = _LiveStreamRegistry()
-
-
-@dataclass
-class _StreamingLogAccumulator:
-    """
-    Line-buffered SSE parser for request logging.
-
-    Network chunk boundaries are not aligned with SSE event boundaries, especially on the
-    logosnode websocket path. Buffer until complete lines are available so streamed usage
-    metadata is not lost when it arrives split across chunks.
-    """
-
-    buffer: str = ""
-    full_text: str = ""
-    first_chunk: Optional[Dict[str, Any]] = None
-    last_chunk: Optional[Dict[str, Any]] = None
-    # Terminal Response object from a Responses-API stream (the
-    # ``response.completed`` / ``response.incomplete`` / ``response.failed``
-    # event carries the full response including usage).
-    responses_final: Optional[Dict[str, Any]] = None
-    _saw_responses_events: bool = False
-    # Usage accumulated from an Anthropic Messages stream. Kept apart from
-    # ``last_chunk`` because that stream ends on ``message_stop``, which carries
-    # no usage and would otherwise erase the figures that arrived one event
-    # earlier — see _consume_messages_event.
-    messages_usage: Optional[Dict[str, Any]] = None
-    _saw_messages_events: bool = False
-    # True once message_delta's usage has arrived — the only event whose
-    # output_tokens is settled. message_start's is a one-token placeholder
-    # that must not stand in for the running count; see streamed_tokens.
-    _messages_final_usage: bool = False
-    # The client may close immediately after receiving the protocol's terminal
-    # event. Treat that as a completed response, even if the worker transport's
-    # following stream_end frame has not been consumed yet.
-    terminal_event_received: bool = False
-    # An upstream ``data: {"error": {...}}`` frame delivered mid-stream with an
-    # HTTP 200 (a content filter, context-length, or rate limit that fired after
-    # generation began). The bytes are forwarded to the client, but the request
-    # is not a success.
-    upstream_error: Optional[Dict[str, Any]] = None
-    # Text deltas seen so far. The exact completion count only arrives with the
-    # terminal usage event, which is no help to anyone watching the request run
-    # — so the delta count stands in for it until then. See streamed_tokens.
-    delta_count: int = 0
-    _decoder: Any = field(
-        default_factory=lambda: codecs.getincrementaldecoder("utf-8")(errors="replace"),
-        repr=False,
-    )
-
-    def feed(self, chunk: bytes | str) -> None:
-        if isinstance(chunk, bytes):
-            text = self._decoder.decode(chunk, final=False)
-        else:
-            text = self._decoder.decode(b"", final=True) + str(chunk)
-            self._decoder.reset()
-        self.buffer += text
-        self._consume_complete_lines()
-
-    def finish(self) -> None:
-        self.buffer += self._decoder.decode(b"", final=True)
-        self._decoder.reset()
-        if not self.buffer:
-            return
-        remainder = self.buffer
-        self.buffer = ""
-        for line in remainder.splitlines():
-            self._consume_line(line.rstrip("\r"))
-
-    def streamed_tokens(self) -> Dict[str, int]:
-        """Best current view of this request's tokens, mid-stream.
-
-        Prompt tokens are exact as soon as the stream opens — every surface
-        states them up front. Completion tokens are not: the count settles only
-        in the terminal usage event, which is precisely the moment a live view
-        stops being interesting. Until it arrives the text deltas are counted
-        instead, one apiece.
-
-        That is an approximation, and it is the right one to make: vLLM emits a
-        delta per token, so it tracks the real figure closely, and speculative
-        decoding is the case where it can undercount. It is only ever shown
-        while the request is running — the settled usage replaces it on
-        completion, so nothing is stored or billed from this.
-        """
-        usage = self.usage()
-        prompt = usage.get("prompt_tokens") or usage.get("input_tokens") or 0
-        if not isinstance(prompt, int):
-            prompt = 0
-        # A Messages stream's message_start already carries an
-        # ``output_tokens`` figure — but it is a one-token placeholder, not a
-        # count, and trusting it pins the live figure at 1 for the whole
-        # generation. Until message_delta settles it, the deltas are the count.
-        if self._saw_messages_events and not self._messages_final_usage:
-            return {"prompt_tokens": prompt, "completion_tokens": self.delta_count}
-        completion = usage.get("completion_tokens") or usage.get("output_tokens") or 0
-        if not isinstance(completion, int) or completion <= 0:
-            completion = self.delta_count
-        return {"prompt_tokens": prompt, "completion_tokens": completion}
-
-    def usage(self) -> Dict[str, Any]:
-        if isinstance(self.responses_final, dict):
-            usage = self.responses_final.get("usage")
-            if isinstance(usage, dict):
-                return usage
-        if isinstance(self.messages_usage, dict) and self.messages_usage:
-            return self.messages_usage
-        if isinstance(self.last_chunk, dict):
-            usage = self.last_chunk.get("usage")
-            if isinstance(usage, dict):
-                return usage
-        return {}
-
-    def response_payload(self) -> Dict[str, Any]:
-        # Responses-API stream: the terminal event already carries the complete
-        # response (output items + usage) — log it verbatim. If the stream was
-        # cut off before the terminal event, fall back to the accumulated text.
-        if isinstance(self.responses_final, dict):
-            return self.responses_final
-        if self._saw_responses_events:
-            return {"content": self.full_text}
-        if self._saw_messages_events:
-            # No chunk to rebuild from: an Anthropic stream never sends the
-            # response as one object, only the events that assemble it.
-            payload: Dict[str, Any] = {"content": self.full_text}
-            if self.messages_usage:
-                payload["usage"] = self.messages_usage
-            return payload
-
-        usage = self.usage()
-        response_payload: Dict[str, Any] = {"content": self.full_text}
-        base_payload = None
-
-        if self.first_chunk:
-            base_payload = self.first_chunk.copy()
-        if self.last_chunk:
-            if base_payload is None:
-                base_payload = self.last_chunk.copy()
-            else:
-                for key, value in self.last_chunk.items():
-                    if key not in base_payload:
-                        base_payload[key] = value
-
-        if base_payload:
-            response_payload = base_payload
-            if "choices" in response_payload and response_payload["choices"]:
-                response_payload["choices"][0]["delta"] = {"content": self.full_text}
-        if usage:
-            response_payload["usage"] = usage
-        return response_payload
-
-    def _consume_complete_lines(self) -> None:
-        while "\n" in self.buffer:
-            line, self.buffer = self.buffer.split("\n", 1)
-            self._consume_line(line.rstrip("\r"))
-
-    def _consume_line(self, line: str) -> None:
-        stripped = line.strip()
-        if not stripped:
-            return
-        if stripped == "data: [DONE]":
-            self.terminal_event_received = True
-            return
-        if not stripped.startswith("data: "):
-            return
-        try:
-            blob = json.loads(stripped[6:])
-        except json.JSONDecodeError:
-            return
-        if not isinstance(blob, dict):
-            return
-
-        blob_error = blob.get("error")
-        if isinstance(blob_error, dict) and blob_error:
-            self.upstream_error = blob_error
-            return
-        if isinstance(blob_error, str) and blob_error:
-            self.upstream_error = {"message": blob_error}
-            return
-
-        event_type = blob.get("type")
-        if isinstance(event_type, str) and event_type.startswith("response."):
-            if event_type in {"response.completed", "response.incomplete", "response.failed"}:
-                self.terminal_event_received = True
-            self._consume_responses_event(event_type, blob)
-            return
-        if isinstance(event_type, str) and event_type in _MESSAGES_EVENT_TYPES:
-            if event_type == "message_stop":
-                self.terminal_event_received = True
-            self._consume_messages_event(event_type, blob)
-            return
-
-        self.last_chunk = blob
-        if self.first_chunk is None:
-            self.first_chunk = blob
-
-        choices = blob.get("choices")
-        if isinstance(choices, list) and choices:
-            delta = choices[0].get("delta", {})
-            if isinstance(delta, dict):
-                content = delta.get("content", "")
-                if content:
-                    self.full_text += content
-                    self.delta_count += 1
-
-    def _consume_responses_event(self, event_type: str, blob: Dict[str, Any]) -> None:
-        """Consume one Responses-API SSE event (``{"type": "response.*", ...}``)."""
-        self._saw_responses_events = True
-        if event_type == "response.output_text.delta":
-            delta = blob.get("delta")
-            if isinstance(delta, str):
-                self.full_text += delta
-                self.delta_count += 1
-        elif event_type in {"response.completed", "response.incomplete", "response.failed"}:
-            response = blob.get("response")
-            if isinstance(response, dict):
-                self.responses_final = response
-            # A terminal ``response.failed`` is the Responses-API equivalent of a
-            # mid-stream ``data: {"error": {...}}`` frame: generation began, the
-            # bytes reached the client, but the request is not a success and must
-            # not be billed. ``response.incomplete`` is deliberately excluded: it
-            # means the generation stopped at max_output_tokens or a content
-            # filter, and the partial output it carries is real, billed usage.
-            if event_type == "response.failed":
-                err = response.get("error") if isinstance(response, dict) else None
-                if isinstance(err, dict) and err:
-                    self.upstream_error = err
-                elif isinstance(err, str) and err:
-                    self.upstream_error = {"message": err}
-                else:
-                    self.upstream_error = {"message": "Responses API stream reported response.failed"}
-
-    def _consume_messages_event(self, event_type: str, blob: Dict[str, Any]) -> None:
-        """Consume one Anthropic Messages SSE event.
-
-        This is what Claude Code speaks, so it covers every session set up
-        through the AI-tools page — and none of it was being counted. The usage
-        arrives in two places and neither is the last event:
-
-            message_start   {"message": {"usage": {"input_tokens": 14, …}}}
-            content_block_delta …
-            message_delta   {"usage": {"input_tokens": 14, "output_tokens": 2}}
-            message_stop    — nothing
-
-        Reading the final chunk's ``usage`` therefore found ``message_stop`` and
-        came away empty, so every such request was logged with no tokens at all.
-        Both events are merged instead, later winning, since message_delta
-        carries the settled figures.
-        """
-        self._saw_messages_events = True
-        if event_type == "content_block_delta":
-            delta = blob.get("delta")
-            if isinstance(delta, dict):
-                text = delta.get("text")
-                if isinstance(text, str):
-                    self.full_text += text
-                    self.delta_count += 1
-            return
-
-        usage: Any = None
-        if event_type == "message_start":
-            message = blob.get("message")
-            if isinstance(message, dict):
-                usage = message.get("usage")
-        elif event_type == "message_delta":
-            usage = blob.get("usage")
-        if not isinstance(usage, dict):
-            return
-
-        if event_type == "message_delta":
-            self._messages_final_usage = True
-        merged = dict(self.messages_usage or {})
-        merged.update(usage)
-        # Anthropic omits a total — it is the sum, so it says it once. Logos
-        # stores one row per token type and the statistics page reads
-        # total_tokens, so a stream without it lands as zero tokens used.
-        prompt = merged.get("input_tokens")
-        completion = merged.get("output_tokens")
-        if isinstance(prompt, int) and isinstance(completion, int):
-            merged["total_tokens"] = prompt + completion
-        self.messages_usage = merged
-
-
-def _usage_tokens_from_payload(
-    response_payload: Any,
-    request_payload: Any = None,
-    path: str = "",
-    billable_request: bool = True,
-) -> Dict[str, int]:
-    """Canonical token usage for a response, merged with the derived non-token
-    billable quantities (``billed_requests``, characters, images, ...) so the
-    stored ``usage_tokens`` rows carry everything ``logos_price_usage`` prices."""
-    usage, _ = finalize_billing_inputs(request_payload, response_payload, path)
-    if not billable_request:
-        usage.pop("billed_requests", None)
-    return usage
 
 
 # One currency unit = 100 cents = 1e8 micro-cents. The unit is USD, not EUR:
@@ -1993,37 +880,6 @@ app.add_middleware(
 )
 
 
-class APIPrefixStripperMiddleware:
-    """Strip a leading `/api` from incoming HTTP and WebSocket paths.
-
-    Lets the UI namespace all of its calls under `/api/*` so a single Traefik
-    route handles every UI-internal endpoint, while existing bare paths
-    (workernode protocol, public `/v1`/`/openai`/`/jobs`, `/docs`, `/metrics`)
-    still resolve unchanged.
-    """
-
-    def __init__(self, app, prefix: str = "/api") -> None:
-        self.app = app
-        self.prefix = prefix.rstrip("/")
-        self._prefix_bytes = self.prefix.encode("ascii")
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] in ("http", "websocket"):
-            path = scope.get("path", "")
-            if path == self.prefix or path.startswith(self.prefix + "/"):
-                new_scope = dict(scope)
-                new_scope["path"] = path[len(self.prefix) :] or "/"
-                raw = scope.get("raw_path")
-                if isinstance(raw, bytes) and raw.startswith(self._prefix_bytes):
-                    rest = raw[len(self._prefix_bytes) :]
-                    q = rest.find(b"?")
-                    path_part = rest[:q] if q >= 0 else rest
-                    query_part = rest[q:] if q >= 0 else b""
-                    new_scope["raw_path"] = (path_part or b"/") + query_part
-                scope = new_scope
-        await self.app(scope, receive, send)
-
-
 app.add_middleware(APIPrefixStripperMiddleware, prefix="/api")
 
 
@@ -2078,909 +934,6 @@ async def _generic_exception_handler(request: Request, exc: Exception) -> JSONRe
     """Catch-all: return HTTP 500 without leaking stack traces or internal details."""
     logger.exception("Unhandled exception on %s", request.url.path)
     return openai_error_response(500, "Internal server error")
-
-
-_HEALTH_STATUS_RANK: Dict[str, int] = {"DOWN": 0, "DEGRADED": 1, "UP": 2}
-
-
-def _model_deployment_status(
-    deployment: Dict[str, Any], provider_id: int, worker_ids: set[int], snapshots: Dict[int, Dict[str, Any]]
-) -> str:
-    """Serveability of one model deployment for the per-model health breakdown.
-
-    Deployments outside the local provider inventory are cloud — UP whenever
-    configured, because Logos does not probe cloud providers (matching the
-    overall /health behaviour). Worker-backed
-    deployments are UP when the worker is online, the model is calibrated, and
-    a lane is loaded/running or sleeping (a wake is fast). A model that only
-    needs a cold load, or that lives on a worker reporting an unhealthy node,
-    is DEGRADED; a missing/stale worker or missing calibration is DOWN.
-    """
-    if provider_id not in worker_ids:
-        return "UP"
-    snapshot = snapshots.get(provider_id)
-    if snapshot is None:
-        return "DOWN"
-    model_name = str(deployment.get("model_name") or "")
-    if model_name not in (snapshot.get("capabilities_models") or []):
-        return "DOWN"
-    runtime = snapshot.get("runtime") if isinstance(snapshot.get("runtime"), dict) else {}
-    node_health = runtime.get("node_health")
-    if isinstance(node_health, dict) and node_health.get("healthy") is False:
-        return "DEGRADED"
-    states = {
-        str(lane.get("runtime_state") or "").strip()
-        for lane in (runtime.get("lanes") or [])
-        if isinstance(lane, dict) and str(lane.get("model") or "").strip() == model_name
-    }
-    if states & {"loaded", "running", "sleeping"}:
-        return "UP"
-    # No loaded/running/sleeping lane: a full cold load is needed first.
-    return "DEGRADED"
-
-
-@app.get("/health", tags=["monitoring"])
-async def health():
-    """Report overall health plus a breakdown of what is serveable.
-
-    Serving local inference is the core function of the orchestrator, so the
-    overall ``status`` is only UP when there is a live (non-stale heartbeat)
-    local worker that declares at least one capable model. When every local
-    provider is offline (or none expose a capable model) we return 503/DOWN so
-    external monitors surface the degradation instead of seeing a misleading UP.
-
-    The body always breaks the state down per backend so callers can tell what
-    exactly is down: ``local_models`` (logosnode workers) and ``cloud_models``
-    (Azure/cloud deployments). Cloud may still be serveable even while local is
-    down, which the ``detail`` message makes explicit.
-
-    This endpoint stays a lean liveness signal: it is public (its own Traefik
-    router on the secure entrypoint, and liveness probes call it), so it must
-    not carry the model catalogue. The per-model view applications need lives
-    on the secret-gated /internal/model_health endpoint instead.
-    """
-    local_ok = False
-    cloud_ok = False
-    try:
-        with DBManager() as db:
-            inventory = db.list_local_providers()
-            deployments = db.get_all_deployments()
-        worker_ids: set[int] = set()
-        for provider in inventory:
-            provider_id = int(provider.get("provider_id") or 0)
-            if provider_id <= 0:
-                continue
-            worker_ids.add(provider_id)
-            snapshot = _logosnode_registry.peek_runtime_snapshot(provider_id)
-            if not _logosnode_snapshot_is_connected(snapshot):
-                continue
-            # Local is serveable if an online worker declares at least one capable model.
-            if snapshot.get("capabilities_models"):
-                local_ok = True
-        # Cloud is serveable if any deployment lives outside the local provider
-        # inventory — by type alone this would miscount legacy local worker
-        # types (ollama, node, ...) as cloud.
-        cloud_ok = any(int(d.get("provider_id") or 0) not in worker_ids for d in deployments)
-    except Exception:
-        logger.exception("Health check failed to evaluate provider state")
-        local_ok = False
-        cloud_ok = False
-
-    payload = {
-        "status": "UP" if local_ok else "DOWN",
-        "local_models": "UP" if local_ok else "DOWN",
-        "cloud_models": "UP" if cloud_ok else "DOWN",
-    }
-    if not local_ok:
-        payload["detail"] = "No local provider with a capable model is online." + (
-            " Cloud models may still be served." if cloud_ok else " No cloud models are configured."
-        )
-    return JSONResponse(status_code=200 if local_ok else 503, content=payload)
-
-
-@app.get("/internal/model_health", tags=["admin"])
-async def internal_model_health(request: Request):
-    """Current per-model health, for the Spring webservice.
-
-    Lane state, worker connection state, and node health exist only in the
-    orchestrator's worker registry, so the webservice's get_model_health
-    endpoint (API-key authenticated, permission filtered) is served from this
-    payload. It is secret-gated rather than part of /health because /health is
-    public: the full model catalogue must not be readable without a credential.
-
-    Status codes mirror /health — 503 while every local worker is down, whose
-    body still carries the breakdown (cloud models may be serveable) — and the
-    entries expose only model names and statuses, best across deployments
-    (see :func:`_model_deployment_status`).
-    """
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal model health endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-    local_ok = False
-    model_status: Dict[str, str] = {}
-    try:
-        with DBManager() as db:
-            inventory = db.list_local_providers()
-            deployments = db.get_all_deployments_with_names()
-        worker_ids: set[int] = set()
-        snapshots: Dict[int, Dict[str, Any]] = {}
-        for provider in inventory:
-            provider_id = int(provider.get("provider_id") or 0)
-            if provider_id <= 0:
-                continue
-            worker_ids.add(provider_id)
-            snapshot = _logosnode_registry.peek_runtime_snapshot(provider_id)
-            if not _logosnode_snapshot_is_connected(snapshot):
-                continue
-            snapshots[provider_id] = snapshot
-            if snapshot.get("capabilities_models"):
-                local_ok = True
-        for deployment in deployments:
-            model_name = str(deployment.get("model_name") or "").strip()
-            if not model_name:
-                continue
-            status = _model_deployment_status(
-                deployment, int(deployment.get("provider_id") or 0), worker_ids, snapshots
-            )
-            current = model_status.get(model_name)
-            if current is None or _HEALTH_STATUS_RANK[status] > _HEALTH_STATUS_RANK[current]:
-                model_status[model_name] = status
-    except Exception:
-        logger.exception("Model health check failed to evaluate provider state")
-        local_ok = False
-        model_status = {}
-
-    return JSONResponse(
-        status_code=200 if local_ok else 503,
-        content={"models": [{"name": name, "status": model_status[name]} for name in sorted(model_status)]},
-    )
-
-
-@app.get("/metrics", tags=["monitoring"])
-async def prometheus_metrics(request: Request):
-    """Prometheus metrics endpoint. Requires PROMETHEUS_API_KEY env var to be set.
-    Pass the key via `Authorization: Bearer <key>` header."""
-    if not _PROMETHEUS_API_KEY:
-        raise HTTPException(
-            status_code=403,
-            detail="Metrics endpoint disabled (PROMETHEUS_API_KEY not configured)",
-        )
-    auth = request.headers.get("authorization", "")
-    token = auth.removeprefix("Bearer ").strip() if auth.lower().startswith("bearer ") else auth.strip()
-    if not hmac.compare_digest(token.encode("utf-8"), _PROMETHEUS_API_KEY.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing metrics API key")
-    body, content_type = _prometheus_metrics_response()
-    from starlette.responses import Response
-
-    return Response(content=body, media_type=content_type)
-
-
-class _RefreshPipelineRequest(BaseModel):
-    rebuild_classifier: bool = False
-
-
-@app.post("/internal/refresh_pipeline", tags=["admin"])
-async def internal_refresh_pipeline(data: _RefreshPipelineRequest, request: Request):
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal refresh endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-    if not _pipeline or not _logosnode_facade or not _azure_facade:
-        raise HTTPException(status_code=503, detail="Pipeline not initialized")
-    logger.info("Pipeline refresh requested by Spring (rebuildClassifier=%s)", data.rebuild_classifier)
-    await refresh_pipeline_runtime_state(rebuild_model_classifier=data.rebuild_classifier)
-    return {"status": "ok"}
-
-
-@app.get("/internal/provider_status", tags=["admin"])
-async def internal_provider_status(request: Request):
-    """Connection state of every local provider, for the Spring webservice.
-
-    The webservice serves the statistics VRAM payload from persisted snapshots
-    only; live connection state (online/offline) exists solely in the
-    orchestrator's worker registry, so it is exposed here for enrichment.
-    """
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal provider status endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-    with DBManager() as db:
-        inventory = db.list_local_providers()
-
-    providers = []
-    for provider in inventory:
-        provider_id = int(provider.get("provider_id") or 0)
-        if provider_id <= 0:
-            continue
-        runtime_snapshot = _logosnode_registry.peek_runtime_snapshot(provider_id)
-        connected = _logosnode_snapshot_is_connected(runtime_snapshot)
-        last_heartbeat = runtime_snapshot.get("last_heartbeat") if runtime_snapshot else None
-        if isinstance(last_heartbeat, datetime.datetime):
-            last_heartbeat = last_heartbeat.isoformat()
-        providers.append(
-            {
-                "provider_id": provider_id,
-                "name": provider.get("name"),
-                "provider_type": provider.get("provider_type"),
-                "connected": connected,
-                "connection_state": "online" if connected else "offline",
-                "last_heartbeat": last_heartbeat if isinstance(last_heartbeat, str) else None,
-                "calibrating": _logosnode_registry.is_calibrating(provider_id),
-            }
-        )
-    return {"providers": providers}
-
-
-@app.get("/internal/model_context_windows", tags=["admin"])
-async def internal_model_context_windows(request: Request):
-    """Served context window per model name, for the Spring webservice.
-
-    The effective window lives only in the worker runtime snapshots held by
-    the orchestrator's registry; the webservice enriches its model listings
-    (e.g. the AI-tools setup page) from this map.
-
-    ``windows`` keeps its original shape — model name -> smallest currently
-    served window — so an older webservice keeps working. ``stats`` adds the
-    ``best`` and ``native`` numbers next to it; see
-    :func:`_served_context_window_stats`.
-    """
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-    stats = _served_context_window_stats()
-    return {
-        "windows": {model: entry["current_min"] for model, entry in stats.items() if "current_min" in entry},
-        "stats": stats,
-    }
-
-
-def _require_internal_secret(request: Request) -> None:
-    """Authenticate an /internal/* call: the shared secret, no user context."""
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-
-@app.get("/internal/live_streams", tags=["admin"])
-async def internal_live_streams(request: Request):
-    """Token counts of the requests streaming right now, for the statistics page.
-
-    A finished request's usage is in the database; one still running is only
-    here, in the process the chunks pass through. Without this the request feed
-    shows a row with no numbers for the whole minute a long generation takes,
-    and then the totals appear at once when it ends.
-
-    Cheap by construction: a dict of the in-flight requests, no database work,
-    and the webservice already polls on the cadence it pushes the feed at.
-    """
-    _require_internal_secret(request)
-    return {"streams": _live_streams.snapshot()}
-
-
-# How often the live-stream SSE connection checks for a changed snapshot. The
-# push is event-driven in effect — a token delta bumps the registry's version
-# and the very next check sends it — but the check itself is a short poll so
-# the endpoint stays a plain generator instead of event plumbing.
-_LIVE_STREAMS_SSE_TICK_S = 0.2
-
-
-@app.get("/internal/live_streams/stream", tags=["admin"])
-async def internal_live_streams_stream(request: Request):
-    """The live view as an SSE stream: the current snapshot, then one event
-    per change.
-
-    The statistics page's token counts used to move at the webservice's poll
-    interval; this lets the webservice forward a change the moment it happens
-    and push it to its websocket clients in real time. The payload is the
-    same shape as ``/internal/live_streams`` so both clients parse one way.
-    """
-    _require_internal_secret(request)
-
-    async def events():
-        version = -1
-        while True:
-            current = _live_streams.version
-            if current != version:
-                version = current
-                yield f"data: {json.dumps({'streams': _live_streams.snapshot()})}\n\n"
-            else:
-                # A comment line: keeps idle (but healthy) connections from
-                # being reaped by anything in the middle.
-                yield ": ping\n\n"
-            await asyncio.sleep(_LIVE_STREAMS_SSE_TICK_S)
-
-    return StreamingResponse(
-        events(),
-        media_type="text/event-stream",
-        headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
-    )
-
-
-@app.get("/internal/calibration_probe_logs", tags=["admin"])
-def internal_calibration_probe_logs(model_name: str, request: Request):
-    """Every node's most recent calibration probe log for one model.
-
-    Backs the webservice's model-error-report page (Complete Logs tab) —
-    it resolves a model id to a model name, then asks here for what every
-    provider that has calibrated it reported.
-    """
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-    with DBManager() as db:
-        rows = db.get_calibration_probe_logs_by_model(model_name)
-    return JSONResponse(status_code=200, content=jsonable_encoder({"logs": rows}))
-
-
-@app.get("/internal/calibration_probe_logs/fetch", tags=["admin"])
-async def internal_fetch_calibration_log(provider_id: int, model_name: str, request: Request):
-    """On-demand fetch of a model's full calibration log from the worker.
-
-    Backs the Download-full-logs button: successful runs skip storing
-    log_text, so this re-reads the on-disk log live. Pure file read, no
-    GPU impact; needs the worker online and not yet recalibrated since.
-    """
-    _require_internal_secret(request)
-
-    try:
-        result = await _logosnode_registry.send_command(
-            provider_id,
-            "get_calibration_log",
-            params={"model_name": model_name},
-            timeout_seconds=20,
-        )
-    except LogosNodeOfflineError as exc:
-        return JSONResponse(status_code=503, content={"error": str(exc) or "Worker not connected"})
-    except LogosNodeCommandError as exc:
-        return JSONResponse(status_code=404, content={"error": str(exc)})
-
-    return JSONResponse(status_code=200, content=jsonable_encoder({"log_text": result.get("log_text", "")}))
-
-
-@app.get("/internal/compatibility_precheck", tags=["admin"])
-async def internal_compatibility_precheck(model_name: str, request: Request, provider_id: int | None = None):
-    """On-demand HF compatibility check for one model, across one or all
-    connected logosnode workers.
-
-    Unlike calibration itself, this is safe to call any time — including
-    during the day, with live traffic on the nodes — because the worker-side
-    check (run_compatibility_precheck) never touches vLLM or lanes: it's
-    only an HF metadata fetch plus a read-only nvidia-smi query. Omitting
-    provider_id fans out across every connected worker, answering "would
-    this model run on ANY node" rather than requiring the caller to already
-    know which one to ask.
-    """
-    _require_internal_secret(request)
-
-    provider_ids = [provider_id] if provider_id is not None else _logosnode_registry.active_provider_ids()
-    if not provider_ids:
-        return JSONResponse(status_code=200, content={"model": model_name, "results": []})
-
-    async def _check_one(pid: int) -> dict[str, Any]:
-        pname = _resolve_provider_name(pid)
-        try:
-            result = await _logosnode_registry.send_command(
-                pid, "run_compatibility_precheck", params={"model": model_name}, timeout_seconds=30
-            )
-            # A worker reply with "result": null bypasses send_command's
-            # dict type hint (a present null isn't its .get() default) and
-            # would otherwise raise **result below, outside this try —
-            # taking every other provider's result down with it.
-            if not isinstance(result, dict):
-                return {
-                    "provider_id": pid,
-                    "provider_name": pname,
-                    "ok": False,
-                    "error": f"Worker returned an unexpected response type: {type(result).__name__}",
-                }
-            return {"provider_id": pid, "provider_name": pname, **result}
-        except LogosNodeOfflineError:
-            return {"provider_id": pid, "provider_name": pname, "ok": False, "error": "Worker not connected"}
-        except LogosNodeCommandError as exc:
-            return {"provider_id": pid, "provider_name": pname, "ok": False, "error": str(exc)}
-        except Exception as exc:  # noqa: BLE001
-            # An unexpected failure for one provider must not lose the
-            # results already gathered for every other one — see below.
-            logger.exception("[Precheck] unexpected failure checking provider %s for %s", pid, model_name)
-            return {"provider_id": pid, "provider_name": pname, "ok": False, "error": f"Unexpected error: {exc}"}
-
-    results = await asyncio.gather(*(_check_one(pid) for pid in provider_ids))
-    return JSONResponse(status_code=200, content=jsonable_encoder({"model": model_name, "results": list(results)}))
-
-
-class _InternalCalibrateRequest(BaseModel):
-    provider_id: int
-
-
-class _InternalDeleteLaneRequest(BaseModel):
-    provider_id: int
-    lane_id: str
-
-
-class _InternalAddLaneRequest(BaseModel):
-    provider_id: int
-    lane: dict[str, Any]
-
-
-class _InternalBenchmarkRequest(BaseModel):
-    model_provider_id: int = Field(gt=0)
-    samples: int = Field(default=5, gt=0, le=100)
-    max_output_tokens: int = Field(default=512, gt=0, le=4096)
-
-
-@app.post("/internal/model_benchmarks/run", tags=["admin"])
-async def internal_run_model_benchmark(data: _InternalBenchmarkRequest, request: Request):
-    """Queue a fixed GSM8K GuideLLM run for one exact provider-model pair."""
-    _require_internal_secret(request)
-
-    # One orchestrator process owns benchmark execution. Serialize the short
-    # check-and-create section in memory so simultaneous starts cannot both
-    # admit a benchmark for the same provider.
-    with _benchmark_admission_lock, DBManager() as db:
-        target = db.get_model_provider_benchmark_target(data.model_provider_id)
-        if target is None:
-            raise HTTPException(status_code=404, detail="Provider-model pair not found")
-        provider_id = int(target["provider_id"])
-        provider_type = _normalize_provider_type(str(target.get("provider_type") or ""))
-        endpoint = str(target.get("target") or "").strip()
-        if provider_type != "logosnode" and not endpoint.startswith(("http://", "https://")):
-            raise HTTPException(status_code=409, detail="Provider-model pair has no valid endpoint")
-        api_key = str(target.get("api_key") or "").strip()
-        if provider_type != "logosnode" and api_key and not credential_transport_is_secure(endpoint):
-            raise HTTPException(
-                status_code=409,
-                detail="Provider credentials require HTTPS (plain HTTP is allowed only on loopback)",
-            )
-
-        active = db.find_active_model_benchmark_job(provider_id)
-        if active is not None:
-            return JSONResponse(
-                status_code=409,
-                content=jsonable_encoder(
-                    {
-                        "error": f"A benchmark is already running on {target['provider_name']}",
-                        "job_id": active["id"],
-                        "status": active["status"],
-                    }
-                ),
-            )
-        runtime_snapshot = _logosnode_registry.peek_runtime_snapshot(provider_id)
-        if provider_type == "logosnode":
-            if runtime_snapshot is None or not _logosnode_snapshot_is_connected(runtime_snapshot):
-                raise HTTPException(status_code=503, detail=f"Provider {target['provider_name']} is offline")
-            if not runtime_snapshot.get("first_status_received"):
-                raise HTTPException(status_code=503, detail="Provider has not sent its first status yet")
-
-        model_name = str(target["model_name"])
-        serving_configuration = extract_serving_configuration(runtime_snapshot, model_name)
-        job_payload = {
-            "model_provider_id": data.model_provider_id,
-            "provider_id": provider_id,
-            "provider_name": target["provider_name"],
-            "model_id": target["model_id"],
-            "model_name": model_name,
-            "dataset": BENCHMARK_DATASET,
-            "subset": "main",
-            "split": "test",
-            "samples": data.samples,
-            "max_output_tokens": data.max_output_tokens,
-            "provider_session_id": runtime_snapshot.get("session_id") if runtime_snapshot else None,
-        }
-        job_id = db.create_job_record(
-            payload=job_payload,
-            api_key_id=None,
-            team_id=None,
-            user_id=None,
-            environment="model-provider-benchmark",
-        )
-
-    is_internal_worker_benchmark = provider_type == "logosnode"
-    benchmark_target = (
-        internal_benchmark_target(job_id) if is_internal_worker_benchmark else resolve_benchmark_target(endpoint)
-    )
-    request_headers = (
-        benchmark_affinity_headers(
-            secret=_INTERNAL_SECRET,
-            job_id=job_id,
-            provider_id=provider_id,
-            model=model_name,
-        )
-        if is_internal_worker_benchmark
-        else None
-    )
-    worker_preparer = (
-        (lambda: _capacity_planner.prepare_benchmark_lane(provider_id, model_name))
-        if request_headers is not None and _capacity_planner is not None
-        else None
-    )
-
-    task = asyncio.create_task(
-        run_benchmark_job(
-            job_id=job_id,
-            model_provider_id=data.model_provider_id,
-            target=benchmark_target,
-            model=model_name,
-            api_key=None if is_internal_worker_benchmark else api_key or None,
-            samples=data.samples,
-            max_output_tokens=data.max_output_tokens,
-            serving_configuration=serving_configuration,
-            serving_configuration_getter=lambda: extract_serving_configuration(
-                _logosnode_registry.peek_runtime_snapshot(provider_id), model_name
-            ),
-            request_headers=request_headers,
-            worker_preparer=worker_preparer,
-            worker_session_is_current=(
-                (
-                    lambda: (_logosnode_registry.peek_runtime_snapshot(provider_id) or {}).get("session_id")
-                    == job_payload["provider_session_id"]
-                )
-                if is_internal_worker_benchmark
-                else None
-            ),
-        )
-    )
-    _background_tasks.add(task)
-    _benchmark_tasks.add(task)
-    _benchmark_tasks_by_job[job_id] = task
-    if is_internal_worker_benchmark:
-        _benchmark_sessions_by_job[job_id] = (provider_id, str(job_payload["provider_session_id"]))
-    task.add_done_callback(lambda done, jid=job_id: _forget_benchmark_task(jid, done))
-    return JSONResponse(
-        status_code=202,
-        content={
-            "job_id": job_id,
-            "status": "pending",
-            "provider_id": provider_id,
-            "provider_name": target["provider_name"],
-            "model_provider_id": data.model_provider_id,
-            "model_name": model_name,
-        },
-    )
-
-
-@app.post("/internal/model_benchmarks/jobs/{job_id}/cancel", tags=["admin"])
-async def internal_cancel_model_benchmark(job_id: int, request: Request):
-    """Cancel one active benchmark and release its lease."""
-    _require_internal_secret(request)
-    if not _cancel_benchmark_job(job_id, "Benchmark cancelled by administrator"):
-        raise HTTPException(status_code=404, detail="Active benchmark job not found")
-    return {"job_id": job_id, "status": "failed"}
-
-
-@app.post("/internal/model_benchmarks/jobs/{job_id}/v1/{path:path}", tags=["admin"])
-async def internal_model_benchmark_completion(job_id: int, path: str, request: Request):
-    """Execute one signed benchmark request without a user API key."""
-    if path.strip("/") != "chat/completions":
-        raise HTTPException(status_code=404, detail="Unsupported benchmark operation")
-
-    try:
-        body = await request.json()
-    except json.JSONDecodeError as exc:
-        raise HTTPException(status_code=400, detail="Invalid JSON body") from exc
-    if not isinstance(body, dict):
-        raise HTTPException(status_code=400, detail="JSON payload must be an object")
-
-    headers = dict(request.headers)
-    header_job_id = headers.get(BENCHMARK_JOB_HEADER)
-    if header_job_id != str(job_id):
-        raise HTTPException(status_code=401, detail="Invalid benchmark worker affinity")
-
-    with DBManager() as db:
-        job = db.get_job(job_id)
-        payload = job.get("request_payload") if job else None
-        if isinstance(payload, str):
-            try:
-                payload = json.loads(payload)
-            except json.JSONDecodeError as exc:
-                raise HTTPException(status_code=401, detail="Invalid benchmark worker affinity") from exc
-        if not isinstance(payload, dict):
-            raise HTTPException(status_code=401, detail="Invalid benchmark worker affinity")
-        try:
-            model_provider_id = int(payload["model_provider_id"])
-        except (KeyError, TypeError, ValueError) as exc:
-            raise HTTPException(status_code=401, detail="Invalid benchmark worker affinity") from exc
-        target = db.get_model_provider_benchmark_target(model_provider_id)
-
-    if target is None or _normalize_provider_type(str(target.get("provider_type") or "")) != "logosnode":
-        raise HTTPException(status_code=404, detail="Benchmark worker deployment no longer exists")
-
-    raw_deployments: list[Deployment] = [
-        {
-            "model_id": int(target["model_id"]),
-            "provider_id": int(target["provider_id"]),
-            "type": "logosnode",
-            "privacy_level": target.get("privacy_level"),
-            "cloud_provider_type": target.get("cloud_provider_type"),
-            "base_url": target.get("base_url"),
-        }
-    ]
-    required_provider_id = _benchmark_provider_affinity(headers, body, raw_deployments)
-    deployments = await _filter_logosnode_deployments(raw_deployments, payload=body)
-    if not deployments:
-        raise HTTPException(status_code=503, detail="Selected benchmark worker is not serving the model")
-
-    # This context is created only after the active job's HMAC and exact
-    # provider-model pair have been validated above. It intentionally has no
-    # user/team limits or billing identity: benchmarks are Logos-internal work.
-    auth = AuthContext(
-        key_value="",
-        api_key_id=-job_id,
-        api_key_name="internal-model-benchmark",
-        key_type="internal",
-        team_id=None,
-        user_id=None,
-        environment="model-provider-benchmark",
-        log_level="FULL",
-        settings={},
-        default_priority=1,
-    )
-    request_id = secrets.token_urlsafe(16)
-    with DBManager() as db:
-        log_result, _ = db.log_usage(
-            api_key_id=None,
-            team_id=None,
-            user_id=None,
-            environment=auth.environment,
-            log_level=auth.log_level,
-            request_id=request_id,
-        )
-    log_id = int(log_result["log-id"])
-    return await _execute_cancelling_on_disconnect(
-        request,
-        deployments=deployments,
-        body=body,
-        headers=headers,
-        auth=auth,
-        path=f"v1/{path.strip('/')}",
-        log_id=log_id,
-        request_id=request_id,
-        required_provider_id=required_provider_id,
-    )
-
-
-class _InternalSleepLaneRequest(BaseModel):
-    provider_id: int
-    lane_id: str
-
-
-class _InternalWakeLaneRequest(BaseModel):
-    provider_id: int
-    lane_id: str
-
-
-@app.post("/internal/logosnode/calibrate_uncalibrated", tags=["admin"])
-async def internal_logosnode_calibrate_uncalibrated(data: _InternalCalibrateRequest, request: Request):
-    """Calibrate uncalibrated models on a worker, called by Spring after JWT validation."""
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-    snap = _logosnode_registry.peek_runtime_snapshot(data.provider_id)
-    if snap is None:
-        return JSONResponse(status_code=503, content={"error": "Worker not connected"})
-    if not snap.get("first_status_received"):
-        return JSONResponse(
-            status_code=503,
-            content={"error": "Worker has not sent its first status yet"},
-        )
-    models = _find_uncalibrated_models_on_provider(data.provider_id)
-    if not models:
-        return {"message": "No uncalibrated models on this worker", "count": 0, "models": []}
-    sleep_level = _calibration_orchestrator._config.sleep_level if _calibration_orchestrator is not None else 1
-    pname = _resolve_provider_name(data.provider_id)
-    try:
-        await _logosnode_registry.send_command(
-            data.provider_id,
-            "start_calibration_session",
-            params={"sleep_level": sleep_level},
-            timeout_seconds=30,
-        )
-    except LogosNodeOfflineError as exc:
-        logger.warning("Internal calibrate-uncalibrated: provider=%s offline: %s", pname, exc)
-        return JSONResponse(status_code=503, content={"error": "Worker not connected"})
-    except LogosNodeCommandError as exc:
-        logger.warning(
-            "Internal calibrate-uncalibrated: start_calibration_session refused on provider=%s: %s", pname, exc
-        )
-        return JSONResponse(status_code=409, content={"error": str(exc)})
-    logger.info(
-        "Internal calibrate-uncalibrated: session started on provider=%s (%d candidate model(s))", pname, len(models)
-    )
-    return {
-        "message": f"Calibration session started on {pname} ({len(models)} candidate model(s))",
-        "count": len(models),
-        "models": models,
-    }
-
-
-@app.post("/internal/logosnode/lanes/delete", tags=["admin"])
-async def internal_logosnode_delete_lane(data: _InternalDeleteLaneRequest, request: Request):
-    """Unload a lane on a worker, called by Spring after JWT validation."""
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="delete_lane",
-        params={"lane_id": data.lane_id},
-    )
-
-
-@app.post("/internal/logosnode/lanes/add", tags=["admin"])
-async def internal_logosnode_add_lane(data: _InternalAddLaneRequest, request: Request):
-    """Manually load a single lane on a worker, called by Spring after JWT validation."""
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-    model = str(data.lane.get("model") or "").strip()
-    if not model:
-        raise HTTPException(status_code=400, detail="lane.model is required")
-
-    if _capacity_planner is None:
-        raise HTTPException(status_code=503, detail="Capacity planner not ready")
-
-    # Answer a refusal synchronously — a background task has nobody to report to.
-    rejection = _capacity_planner.manual_load_rejection_reason(data.provider_id)
-    if rejection is not None:
-        raise HTTPException(status_code=409, detail=rejection)
-
-    # Loading a model takes minutes (the planner budgets 1800 s for the command),
-    # far beyond any caller's HTTP read timeout, and holding a servlet thread
-    # open that long per load is its own problem. So kick it off and return: the
-    # lane appears in the lane-status stream the statistics page already
-    # subscribes to, which is where the operator watches it come up.
-    task = asyncio.create_task(_capacity_planner.load_lane_manually(data.provider_id, model))
-    _background_tasks.add(task)
-    task.add_done_callback(_background_tasks.discard)
-    # 202, not 200: the lane is not loaded when this returns, only scheduled.
-    return JSONResponse(
-        status_code=202,
-        content={"status": "accepted", "model": model, "provider_id": data.provider_id},
-    )
-
-
-@app.post("/internal/logosnode/lanes/sleep", tags=["admin"])
-async def internal_logosnode_sleep_lane(data: _InternalSleepLaneRequest, request: Request):
-    """Sleep a lane on a worker, called by Spring after JWT validation.
-
-    Level 1 keeps the weights resident in host memory, so the lane wakes in
-    seconds; level 2 would release them and pay for a full reload on the next
-    wake — a choice most operators cannot make well, so the button does not
-    offer it and neither does this endpoint.
-
-    In-flight requests are not refused: mode="wait" makes the worker drain
-    them first and only sleep once the lane is idle (a request admitted
-    between drain and sleep makes the worker skip the sleep and stay awake).
-    The command therefore takes as long as the drain — the dispatch budget
-    must cover it, which is why sleep_lane gets the same 120 s as the
-    planner's own sleep commands.
-    """
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-
-    snap = _logosnode_registry.peek_runtime_snapshot(data.provider_id)
-    if snap is None:
-        return JSONResponse(status_code=503, content={"error": "Worker not connected"})
-    if not snap.get("first_status_received"):
-        return JSONResponse(
-            status_code=503,
-            content={"error": "Worker has not sent its first status yet"},
-        )
-    lanes = (snap.get("runtime") or {}).get("lanes") or []
-    lane = next(
-        (item for item in lanes if isinstance(item, dict) and str(item.get("lane_id", "")) == data.lane_id),
-        None,
-    )
-    if lane is None:
-        return JSONResponse(status_code=404, content={"error": f"Lane '{data.lane_id}' not found on this worker"})
-    # "unsupported" is the worker's resolved answer for "cannot sleep this
-    # lane": enable_sleep_mode off for its model (per-model override or the
-    # node-wide kill switch) or a non-vLLM lane. Dispatching would burn the
-    # command budget to fail on the worker, so refuse synchronously with the
-    # reason the panel can display.
-    if str(lane.get("sleep_state", "")).strip().lower() == "unsupported":
-        raise HTTPException(
-            status_code=409,
-            detail="Lane does not support sleep mode; its model is configured without enable_sleep_mode.",
-        )
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="sleep_lane",
-        params={"lane_id": data.lane_id, "level": 1, "mode": "wait"},
-    )
-
-
-@app.post("/internal/logosnode/lanes/wake", tags=["admin"])
-async def internal_logosnode_wake_lane(data: _InternalWakeLaneRequest, request: Request):
-    """Wake a sleeping lane on a worker, called by Spring after JWT validation."""
-    if not _INTERNAL_SECRET:
-        raise HTTPException(status_code=403, detail="Internal endpoint disabled")
-    auth_header = request.headers.get("authorization", "")
-    token = (
-        auth_header.removeprefix("Bearer ").strip()
-        if auth_header.lower().startswith("bearer ")
-        else auth_header.strip()
-    )
-    if not hmac.compare_digest(token.encode("utf-8"), _INTERNAL_SECRET.encode("utf-8")):
-        raise HTTPException(status_code=401, detail="Invalid or missing internal secret")
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="wake_lane",
-        params={"lane_id": data.lane_id},
-    )
 
 
 # ============================================================================
@@ -3041,45 +994,8 @@ def _extract_policy(headers: dict, logos_key: str, body: dict):
     return policy if policy else None
 
 
-def _require_root_access(logos_key: str) -> None:
-    with DBManager() as db:
-        require_logos_admin_key(logos_key, db)
-
-
 def _normalize_provider_type(provider_type: str | None) -> str:
     return normalize_provider_type(provider_type)
-
-
-def _logosnode_insecure_dev_mode_enabled() -> bool:
-    raw = os.getenv("LOGOS_NODE_DEV_ALLOW_INSECURE_HTTP", "")
-    return raw.strip().lower() in {"1", "true", "yes", "on"}
-
-
-def _is_tls_request(request: Request) -> bool:
-    if _logosnode_insecure_dev_mode_enabled():
-        return True
-    if request.url.scheme == "https":
-        return True
-    forwarded = request.headers.get("x-forwarded-proto", "")
-    forwarded_values = [item.strip().lower() for item in forwarded.split(",") if item.strip()]
-    return "https" in forwarded_values
-
-
-def _require_tls_request(request: Request) -> None:
-    if not _is_tls_request(request):
-        raise HTTPException(
-            status_code=400,
-            detail="TLS is required for logosnode auth/session endpoints",
-        )
-
-
-def _build_logosnode_ws_url(request: Request, token: str) -> str:
-    _require_tls_request(request)
-    ws_scheme = "ws" if _logosnode_insecure_dev_mode_enabled() else "wss"
-    host = request.headers.get("host", "")
-    if not host:
-        raise HTTPException(status_code=400, detail="Missing Host header for websocket URL generation")
-    return f"{ws_scheme}://{host}/logosdb/providers/logosnode/session?token={token}"
 
 
 async def _filter_logosnode_deployments(
@@ -5755,239 +3671,6 @@ async def execute_proxy_job(
         _live_streams.finish(request_id)
 
 
-# ============================================================================
-# LOGOSNODE PROVIDER ENDPOINTS
-# ============================================================================
-
-
-def _is_tls_websocket(websocket: WebSocket) -> bool:
-    if _logosnode_insecure_dev_mode_enabled():
-        return True
-    if websocket.url.scheme in {"wss", "https"}:
-        return True
-    forwarded = websocket.headers.get("x-forwarded-proto", "")
-    forwarded_values = [item.strip().lower() for item in forwarded.split(",") if item.strip()]
-    return "https" in forwarded_values or "wss" in forwarded_values
-
-
-@app.post("/logosdb/providers/logosnode/register", tags=["logosnode"])
-async def logosnode_register(data: LogosNodeRegisterRequest):
-    """
-    Root-only provider bootstrap endpoint for LogosWorkerNode providers.
-    """
-    _require_root_access(data.logos_key)
-
-    provider_name = (data.provider_name or "").strip()
-    if not provider_name:
-        raise HTTPException(status_code=400, detail="provider_name is required")
-
-    shared_key = secrets.token_urlsafe(48)
-    with DBManager() as db:
-        result, code = db.add_provider(
-            logos_key=data.logos_key,
-            provider_name=provider_name,
-            base_url=(data.base_url or "").strip(),
-            api_key=shared_key,
-            auth_name="",
-            auth_format="{}",
-            provider_type="logosnode",
-        )
-
-    if code != 200:
-        return JSONResponse(status_code=code, content=result)
-
-    provider_id = result.get("provider-id")
-
-    # Create logosnode_provider_keys entry so deployment queries work
-    try:
-        with DBManager() as db:
-            db.sync_logosnode_capabilities(provider_id, [])
-    except Exception:
-        logger.exception("Failed to create logosnode_provider_keys for provider %s", provider_name)
-
-    return {
-        "provider_id": provider_id,
-        "provider_name": provider_name,
-        "provider_type": "logosnode",
-        "shared_key": shared_key,
-    }
-
-
-@app.post("/logosdb/providers/logosnode/auth", tags=["logosnode"])
-async def logosnode_auth(data: LogosNodeAuthRequest, request: Request):
-    """
-    Authenticate a LogosWorkerNode by its API key.
-
-    The server resolves the provider from the key. The worker never needs
-    to know or send a provider_id.
-    """
-    _require_tls_request(request)
-    with DBManager() as db:
-        provider = db.get_logosnode_provider_by_api_key(data.shared_key)
-
-    if not provider:
-        raise HTTPException(status_code=404, detail="Provider not found for this API key")
-    provider_type = _normalize_provider_type(provider.get("provider_type"))
-    if provider_type != "logosnode":
-        raise HTTPException(status_code=403, detail="Provider is not configured as logosnode")
-
-    provider_id = provider["id"]
-    worker_id = provider.get("name") or f"worker-{provider_id}"
-
-    conflicting_session = await _logosnode_registry.get_conflicting_session(
-        provider_id,
-        worker_id,
-        stale_after_seconds=_LOGOSNODE_STATS_STALE_AFTER_SECONDS,
-    )
-    if conflicting_session is not None:
-        raise HTTPException(
-            status_code=409,
-            detail=(
-                f"Worker '{conflicting_session.worker_id}' is already connected. " f"Stop the existing worker first."
-            ),
-        )
-    token = await _logosnode_registry.issue_ticket(
-        provider_id=provider_id,
-        worker_id=worker_id,
-        capabilities_models=data.capabilities_models,
-        configured_models=data.configured_models or None,
-        ttl_seconds=60,
-    )
-    return {
-        "session_token": token,
-        "ws_url": _build_logosnode_ws_url(request, token),
-        "worker_id": worker_id,
-        "expires_in_seconds": 60,
-    }
-
-
-@app.websocket("/logosdb/providers/logosnode/session")
-async def logosnode_session(websocket: WebSocket, token: str):
-    if not _is_tls_websocket(websocket):
-        await websocket.close(code=1008, reason="TLS required")
-        return
-
-    ticket = await _logosnode_registry.consume_ticket(token)
-    if ticket is None:
-        await websocket.close(code=1008, reason="Invalid or expired token")
-        return
-
-    await websocket.accept()
-    try:
-        session = await _logosnode_registry.attach_session(ticket, websocket)
-    except LogosNodeSessionConflictError as exc:
-        await websocket.close(code=1008, reason=str(exc))
-        return
-    _cancel_benchmarks_for_changed_session(ticket.provider_id, session.session_id)
-
-    try:
-        while True:
-            payload = await websocket.receive_json()
-            if not isinstance(payload, dict):
-                continue
-            msg_type = payload.get("type")
-            if msg_type == "hello":
-                await _logosnode_registry.on_hello(
-                    provider_id=ticket.provider_id,
-                    worker_id=str(payload.get("worker_id", "")).strip() or ticket.worker_id,
-                    capabilities_models=(
-                        payload.get("capabilities_models")
-                        if isinstance(payload.get("capabilities_models"), list)
-                        else None
-                    ),
-                    configured_models=(
-                        payload.get("configured_models") if isinstance(payload.get("configured_models"), list) else None
-                    ),
-                    max_lanes=(
-                        int(payload.get("max_lanes", 0)) if isinstance(payload.get("max_lanes"), (int, float)) else 0
-                    ),
-                    calibrating=(
-                        bool(payload.get("calibrating")) if isinstance(payload.get("calibrating"), bool) else None
-                    ),
-                    actions=(payload.get("actions") if isinstance(payload.get("actions"), list) else None),
-                )
-            elif msg_type == "status":
-                runtime = payload.get("runtime") if isinstance(payload.get("runtime"), dict) else {}
-                await _logosnode_registry.update_runtime(
-                    provider_id=ticket.provider_id,
-                    runtime=runtime,
-                    capabilities_models=(
-                        payload.get("capabilities_models")
-                        if isinstance(payload.get("capabilities_models"), list)
-                        else None
-                    ),
-                    configured_models=(
-                        payload.get("configured_models") if isinstance(payload.get("configured_models"), list) else None
-                    ),
-                    calibrating=(
-                        bool(payload.get("calibrating")) if isinstance(payload.get("calibrating"), bool) else None
-                    ),
-                )
-                _capture_logosnode_provider_snapshot(ticket.provider_id, runtime)
-            elif msg_type == "event":
-                event = payload.get("event") if isinstance(payload.get("event"), dict) else {}
-                await _logosnode_registry.append_event(
-                    provider_id=ticket.provider_id,
-                    event=event,
-                    replay=bool(payload.get("replay", False)),
-                )
-                if event.get("event") == "calibration_probe_log":
-                    try:
-                        await asyncio.to_thread(_capture_calibration_probe_log, ticket.provider_id, event)
-                    except Exception:
-                        logger.debug(
-                            "Failed to persist calibration probe log for provider %s",
-                            _resolve_provider_name(ticket.provider_id),
-                            exc_info=True,
-                        )
-            elif msg_type == "heartbeat":
-                await _logosnode_registry.mark_heartbeat(ticket.provider_id)
-            elif msg_type == "command_result":
-                await _logosnode_registry.on_command_result(ticket.provider_id, payload)
-            elif msg_type == "stream_start":
-                await _logosnode_registry.on_stream_start(ticket.provider_id, payload)
-            elif msg_type == "stream_chunk":
-                await _logosnode_registry.on_stream_chunk(ticket.provider_id, payload)
-            elif msg_type == "stream_end":
-                await _logosnode_registry.on_stream_end(ticket.provider_id, payload)
-    except WebSocketDisconnect:
-        pass
-    finally:
-        await _logosnode_registry.detach_session(ticket.provider_id, websocket)
-        current = _logosnode_registry.peek_runtime_snapshot(ticket.provider_id)
-        _cancel_benchmarks_for_changed_session(
-            ticket.provider_id,
-            str(current["session_id"]) if current else None,
-        )
-
-
-@app.post("/logosdb/providers/logosnode/status", tags=["logosnode"])
-async def logosnode_status(data: LogosNodeStatusRequest):
-    _require_root_access(data.logos_key)
-    try:
-        return await _logosnode_registry.get_runtime_snapshot(data.provider_id)
-    except LogosNodeOfflineError as exc:
-        return JSONResponse(status_code=503, content={"error": str(exc)})
-
-
-@app.post("/logosdb/providers/logosnode/devices", tags=["logosnode"])
-async def logosnode_devices(data: LogosNodeStatusRequest):
-    _require_root_access(data.logos_key)
-    try:
-        return {"devices": await _logosnode_registry.get_devices(data.provider_id)}
-    except LogosNodeOfflineError as exc:
-        return JSONResponse(status_code=503, content={"error": str(exc)})
-
-
-@app.post("/logosdb/providers/logosnode/lanes", tags=["logosnode"])
-async def logosnode_lanes(data: LogosNodeStatusRequest):
-    _require_root_access(data.logos_key)
-    try:
-        return {"lanes": await _logosnode_registry.get_lanes(data.provider_id)}
-    except LogosNodeOfflineError as exc:
-        return JSONResponse(status_code=503, content={"error": str(exc)})
-
-
 _LOGOSNODE_CMD_TIMEOUTS: dict[str, int] = {
     "apply_lanes": 180,
     "reconfigure_lane": 180,
@@ -6013,56 +3696,6 @@ async def _dispatch_logosnode_command(provider_id: int, action: str, params: dic
         return JSONResponse(status_code=503, content={"error": str(exc)})
     except LogosNodeCommandError as exc:
         return JSONResponse(status_code=502, content={"error": str(exc)})
-
-
-@app.post("/logosdb/providers/logosnode/lanes/apply", tags=["logosnode"])
-async def logosnode_apply_lanes(data: LogosNodeApplyLanesRequest):
-    _require_root_access(data.logos_key)
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="apply_lanes",
-        params={"lanes": data.lanes},
-    )
-
-
-@app.post("/logosdb/providers/logosnode/lanes/sleep", tags=["logosnode"])
-async def logosnode_sleep_lane(data: LogosNodeSleepLaneRequest):
-    _require_root_access(data.logos_key)
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="sleep_lane",
-        params={"lane_id": data.lane_id, "level": data.level, "mode": data.mode},
-    )
-
-
-@app.post("/logosdb/providers/logosnode/lanes/wake", tags=["logosnode"])
-async def logosnode_wake_lane(data: LogosNodeWakeLaneRequest):
-    _require_root_access(data.logos_key)
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="wake_lane",
-        params={"lane_id": data.lane_id},
-    )
-
-
-@app.post("/logosdb/providers/logosnode/lanes/delete", tags=["logosnode"])
-async def logosnode_delete_lane(data: LogosNodeDeleteLaneRequest):
-    _require_root_access(data.logos_key)
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="delete_lane",
-        params={"lane_id": data.lane_id},
-    )
-
-
-@app.post("/logosdb/providers/logosnode/lanes/reconfigure", tags=["logosnode"])
-async def logosnode_reconfigure_lane(data: LogosNodeReconfigureLaneRequest):
-    _require_root_access(data.logos_key)
-    return await _dispatch_logosnode_command(
-        provider_id=data.provider_id,
-        action="reconfigure_lane",
-        params={"lane_id": data.lane_id, "updates": data.updates},
-    )
 
 
 def _find_uncalibrated_models_on_provider(provider_id: int) -> list[str]:
@@ -6109,246 +3742,9 @@ def _find_uncalibrated_models_on_provider(provider_id: int) -> list[str]:
     return uncalibrated
 
 
-@app.post("/logosdb/providers/logosnode/calibrate_uncalibrated", tags=["logosnode"])
-async def logosnode_calibrate_uncalibrated(data: LogosNodeStatusRequest):
-    """Kick off a worker-driven calibration session immediately.
-
-    The worker picks which uncalibrated models to run and walks them one at
-    a time, emitting ``calibration_*`` events as each completes. The server
-    no longer chooses models or polls status — the response just confirms
-    the session was started and reports which models the worker will see
-    as uncalibrated right now.
-    """
-    _require_root_access(data.logos_key)
-    snap = _logosnode_registry.peek_runtime_snapshot(data.provider_id)
-    if snap is None:
-        return JSONResponse(status_code=503, content={"error": "Worker not connected"})
-    if not snap.get("first_status_received"):
-        return JSONResponse(
-            status_code=503,
-            content={"error": "Worker has not sent its first status yet"},
-        )
-    models = _find_uncalibrated_models_on_provider(data.provider_id)
-    if not models:
-        return {
-            "message": "No uncalibrated models on this worker",
-            "count": 0,
-            "models": [],
-        }
-    sleep_level = _calibration_orchestrator._config.sleep_level if _calibration_orchestrator is not None else 1
-    pname = _resolve_provider_name(data.provider_id)
-    try:
-        await _logosnode_registry.send_command(
-            data.provider_id,
-            "start_calibration_session",
-            params={"sleep_level": sleep_level},
-            timeout_seconds=30,
-        )
-    except LogosNodeOfflineError as exc:
-        logger.warning("Admin calibrate-uncalibrated: provider=%s offline: %s", pname, exc)
-        return JSONResponse(status_code=503, content={"error": "Worker not connected"})
-    except LogosNodeCommandError as exc:
-        logger.warning(
-            "Admin calibrate-uncalibrated: start_calibration_session refused on provider=%s: %s",
-            pname,
-            exc,
-        )
-        return JSONResponse(status_code=409, content={"error": str(exc)})
-    logger.info(
-        "Admin calibrate-uncalibrated: session started on provider=%s (%d candidate model(s))",
-        pname,
-        len(models),
-    )
-    return {
-        "message": f"Calibration session started on {pname} ({len(models)} candidate model(s))",
-        "count": len(models),
-        "models": models,
-    }
-
-
-# ============================================================================
-# DATABASE MANAGEMENT ENDPOINTS
-# ============================================================================
-
-
-@app.post("/logosdb/update_provider_sdi_config", tags=["admin"])
-async def update_provider_sdi_config(data: UpdateProviderSdiConfigRequest):
-    with DBManager() as db:
-        result = db.update_provider_sdi_config(**data.dict())
-    await refresh_pipeline_runtime_state()
-    return result
-
-
-@app.post("/logosdb/connect_model_provider", tags=["admin"])
-async def connect_model_provider(data: ConnectModelProviderRequest):
-    with DBManager() as db:
-        result = db.connect_model_provider(**data.dict())
-    await refresh_pipeline_runtime_state()
-    return result
-
-
-@app.get("/logosdb/scheduler_state", tags=["admin"])
-async def scheduler_state(request: Request):
-    """
-    Debug endpoint to inspect in-memory scheduler and LogosWorkerNode capacity state.
-    """
-    headers = dict(request.headers)
-    authenticate_api_key(headers)
-
-    if not _pipeline or not _logosnode_facade:
-        return JSONResponse(content={"error": "Scheduler not initialized"}, status_code=503)
-
-    payload = {
-        "queue_total": _pipeline.scheduler.get_total_queue_depth(),
-        "logosnode": _logosnode_facade.debug_state(),
-    }
-    prefix_router = getattr(_pipeline.scheduler, "_prefix_router", None)
-    if prefix_router is not None:
-        payload["prefix_affinity"] = prefix_router.debug_state()
-    return JSONResponse(content=payload, status_code=200)
-
-
-def _today_utc() -> str:
-    from datetime import datetime, timezone
-
-    return datetime.now(timezone.utc).strftime("%Y-%m-%d")
-
-
-@app.post("/logosdb/get_ollama_vram_stats", tags=["admin"])
-async def get_ollama_vram_stats(request: Request):
-    """
-    Return live LogosWorkerNode provider VRAM usage for dashboards.
-
-    Request body:
-    {
-        "day": "2025-01-05",                    # Optional, ignored for runtime-backed stats
-        "bucket_seconds": 5,                    # Optional, ignored for compatibility
-        "after_snapshot_id": 0                  # Optional, return only snapshots with
-                                                # snapshot_id > this (incremental polling)
-    }
-
-    Response:
-    {
-        "providers": [
-            {
-                "url": "http://host.docker.internal:11435",
-                "data": [
-                    {"timestamp": "2025-01-05T10:00:00Z", "vram_mb": 4608},
-                    ...
-                ]
-            }
-        ]
-    }
-    """
-    headers = dict(request.headers)
-    auth = authenticate_api_key(headers)
-    logos_key = auth.key_value
-
-    day = _today_utc()
-    after_snapshot_id = 0
-
-    # Tolerate empty/no-body requests for compatibility with older clients.
-    try:
-        body = await request.json()
-        if isinstance(body, dict):
-            if isinstance(body.get("day"), str) and body.get("day", "").strip():
-                day = body["day"].strip()
-            # Honor the incremental cursor so per-second pollers can fetch only
-            # new snapshots instead of re-receiving the whole day on every call.
-            raw_cursor = body.get("after_snapshot_id")
-            if raw_cursor is not None:
-                try:
-                    after_snapshot_id = max(0, int(raw_cursor))
-                except (TypeError, ValueError):
-                    after_snapshot_id = 0
-    except json.JSONDecodeError:
-        pass
-
-    return JSONResponse(
-        content=_build_live_local_provider_vram_payload(logos_key, day=day, after_snapshot_id=after_snapshot_id),
-        status_code=200,
-    )
-
-
-@app.options("/logosdb/get_ollama_vram_stats", tags=["admin"])
-async def get_ollama_vram_stats_options():
-    """CORS preflight for get_ollama_vram_stats."""
-    return JSONResponse(
-        content={},
-        headers={
-            "Access-Control-Allow-Origin": "*",
-            "Access-Control-Allow-Methods": "POST, OPTIONS",
-            "Access-Control-Allow-Headers": "Content-Type, logos_key",
-        },
-    )
-
-
-@app.get("/forward_host", tags=["admin"])
-async def forward_host(request: Request):
-    forwarded = request.headers.get("X-Forwarded-Host") or request.headers.get("Forwarded")
-    return JSONResponse(content={"host": forwarded})
-
-
 # ============================================================================
 # OPENAI-COMPATIBLE MODEL LISTING
 # ============================================================================
-
-
-def _lane_served_context_window(lane: dict, model_profiles: dict) -> int:
-    """Served context window of one lane in tokens, 0 when unknown.
-
-    Mirrors the worker's --max-model-len precedence for vLLM lanes
-    (vllm_process.py): explicit vllm_config value, then a non-sentinel lane
-    context_length (4096 is the shared lane-schema default, meaning "unset"
-    for vLLM), then the calibrated profile value. Ollama lanes always run at
-    their configured context_length. A vLLM lane where none of these are set
-    lets vLLM pick the model's native maximum, which the worker does not
-    report — such lanes yield 0 rather than a guess.
-    """
-    model = lane.get("model")
-    if not model:
-        return 0
-
-    def _as_len(value) -> int:
-        try:
-            value = int(value)
-        except (TypeError, ValueError):
-            return 0
-        return value if value > 0 else 0
-
-    if not lane.get("vllm"):
-        return _as_len(lane.get("context_length"))
-
-    backend_metrics = lane.get("backend_metrics")
-    explicit = _as_len(backend_metrics.get("max_model_len")) if isinstance(backend_metrics, dict) else 0
-    if explicit:
-        return explicit
-    lane_ctx = _as_len(lane.get("context_length"))
-    if lane_ctx and lane_ctx != 4096:
-        return lane_ctx
-    profile = model_profiles.get(model)
-    if isinstance(profile, dict):
-        return _as_len(profile.get("calibration_max_model_len")) or _as_len(profile.get("max_context_length"))
-    return 0
-
-
-def _profile_native_context_length(profile: dict) -> int:
-    """Largest context window a model could ever be served with here.
-
-    The widest window the model's profile reports — its own architectural
-    limit (``max_context_length``), the ``--max-model-len`` calibration settled
-    on (``calibration_max_model_len``), or the widest point of the calibrated
-    KV sweep — whichever is largest. This is the "all-time maximum" the model
-    offers, as opposed to the window a lane happens to run with right now.
-
-    Reading ``calibration_max_model_len`` matters on its own: a model
-    calibration capped its ``--max-model-len`` to fit the pinned KV budget and
-    recorded no wider KV point, so the calibrated cap is the only context the
-    profile reports. Ignoring it made such a model look context-unknown (and
-    the client fall back to a guessed window) while its worker sat connected
-    and ready to serve it at exactly that width (#829).
-    """
-    return derived_reported_context_length(profile)
 
 
 # The high-water mark only changes when a worker snapshot arrives (a few to a
@@ -6570,439 +3966,23 @@ def _model_context_fields(entry: Optional[dict[str, int]]) -> dict[str, int]:
     return fields
 
 
-@app.get("/v1/models", tags=["user-facing"])
-@app.get("/openai/models", tags=["user-facing"], include_in_schema=False)
-async def list_models(request: Request):
-    """
-    List models accessible to the authenticated user (OpenAI-compatible).
-
-    Also served under /openai/models: the /openai prefix mirrors /v1, and the
-    POST catch-all alias cannot answer this GET.
-
-    Returns an OpenAI-compatible response listing all models the user's
-    current API key has access to (Union of Team models and specific API Key models).
-    Stored aliases of an accessible model are listed as additional model ids
-    right after their model, so logical names (e.g. 'local-most-powerful')
-    can be discovered and used directly in requests.
-
-    Returns:
-        JSONResponse matching the OpenAI GET /v1/models spec.
-    """
-    auth = authenticate_api_key(dict(request.headers))
-
-    with DBManager() as db:
-        models = db.get_models_for_api_key(auth.api_key_id)
-
-    stats = _served_context_window_stats()
-    data = []
-    for model in models:
-        name = model["name"]
-        # Aliases resolve to the same model, so they carry the context-window
-        # fields of the lanes serving it.
-        data.append(
-            {
-                "id": name,
-                "object": "model",
-                "created": _SERVER_START_TIME,
-                "owned_by": "logos",
-                **_model_context_fields(stats.get(name)),
-            }
-        )
-        for alias in model.get("aliases") or []:
-            data.append(
-                {
-                    "id": alias,
-                    "object": "model",
-                    "created": _SERVER_START_TIME,
-                    "owned_by": "logos",
-                    **_model_context_fields(stats.get(name)),
-                }
-            )
-
-    return JSONResponse(content={"object": "list", "data": data})
-
-
-@app.get("/v1/models/{model_id:path}", tags=["user-facing"])
-@app.get("/openai/models/{model_id:path}", tags=["user-facing"], include_in_schema=False)
-async def retrieve_model(model_id: str, request: Request):
-    """
-    Retrieve a single model by name (OpenAI-compatible).
-
-    Verifies the authenticated user has access to the requested model
-    through their combined (Team + API Key) model permissions.
-
-    Params:
-        model_id: The model name (used as the OpenAI-style model id).
-        request: Incoming request.
-
-    Returns:
-        JSONResponse matching the OpenAI GET /v1/models/{model} spec.
-
-    Raises:
-        HTTPException(404): Model not found or user lacks access.
-    """
-    auth = authenticate_api_key(dict(request.headers))
-
-    with DBManager() as db:
-        model = db.get_model_for_api_key(auth.api_key_id, model_id)
-        if not model:
-            models = db.get_models_for_api_key(auth.api_key_id)
-            canonical_model_name = _resolve_requested_model_name(model_id, models)
-            if canonical_model_name is not None:
-                model = next(
-                    (entry for entry in models if entry.get("name") == canonical_model_name),
-                    None,
-                )
-
-    if not model:
-        raise HTTPException(status_code=404, detail="Model not found or access denied")
-
-    stats = _served_context_window_stats()
-    return JSONResponse(
-        content={
-            "id": model["name"],
-            "object": "model",
-            "created": _SERVER_START_TIME,
-            "owned_by": "logos",
-            **_model_context_fields(stats.get(model["name"])),
-        }
-    )
-
-
-def _resolve_accessible_model_name(api_key_id: int, model_id: str) -> Optional[str]:
-    """Canonical name of ``model_id`` if this key may use it, else None.
-
-    Shared by the model endpoints below: they all have to accept the same
-    aliases (stored alternative names, planner-sanitized underscores, case
-    differences) and all have to refuse a model the key has no permission for.
-    """
-    with DBManager() as db:
-        model = db.get_model_for_api_key(api_key_id, model_id)
-        if model:
-            return model["name"]
-        models = db.get_models_for_api_key(api_key_id)
-        return _resolve_requested_model_name(model_id, models)
-
-
-@app.post("/v1/models/{model_id:path}/warmup", tags=["user-facing"])
-@app.post("/openai/models/{model_id:path}/warmup", tags=["user-facing"], include_in_schema=False)
-async def warmup_model(model_id: str, request: Request):
-    """Tell the planner a model is about to be used, and return immediately.
-
-    A coding assistant asks for the model list when it starts and then sits
-    idle while the developer reads the terminal — the first real request lands
-    seconds later, and pays for a cold load it could have overlapped with that
-    pause. This turns the startup into a hint.
-
-    It is a *hint*, not a reservation: it records the same latent demand the
-    scheduler records when classification prefers a model it did not get, and
-    wakes the planner cycle early. The planner still decides what to load using
-    its own fairness rules, so a warmup can never evict a lane that real
-    traffic is using, and a burst of them coalesces into one extra cycle. That
-    is also what keeps it from being a way to make the cluster thrash: the most
-    an authenticated caller can do is raise a model it already has access to
-    slightly in the queue of things worth loading.
-
-    Deliberately not "send a tiny request": that bills the caller, occupies a
-    slot, and returns a completion nobody wanted.
-    """
-    auth = authenticate_api_key(dict(request.headers))
-    model_name = _resolve_accessible_model_name(auth.api_key_id, model_id)
-    if model_name is None:
-        raise HTTPException(status_code=404, detail="Model not found or access denied")
-
-    stats = _served_context_window_stats().get(model_name) or {}
-    # A reported window means some lane is serving the model right now, which is
-    # the closest thing to "ready" this endpoint can answer without asking every
-    # worker. Already-warm models still record the hint: it keeps the model from
-    # decaying out of the planner's demand view while a session is open.
-    already_serving = bool(stats.get("current_min"))
-
-    accepted = False
-    if _demand_tracker is not None:
-        _demand_tracker.record_latent_demand(model_name)
-        accepted = True
-    if _capacity_planner is not None:
-        # announce_upcoming_use, not hint_capacity_needed: the hint only wakes
-        # the cycle early, and the demand increment above cannot survive the
-        # per-cycle decay that runs before the planner evaluates it, so on its
-        # own a warmup would wake a cycle that then decides to do nothing —
-        # which is exactly what it did. The announcement is what lets the
-        # planner cold-load on VRAM that is free anyway.
-        _capacity_planner.announce_upcoming_use(model_name)
-        accepted = True
-
-    logger.info(
-        "Warmup requested for model=%s (already serving: %s, hint accepted: %s)",
-        model_name,
-        already_serving,
-        accepted,
-    )
-    return JSONResponse(
-        status_code=202,
-        content={
-            "model": model_name,
-            "status": "serving" if already_serving else "preparing",
-            "hint_accepted": accepted,
-            **_model_context_fields(stats),
-        },
-    )
-
-
 # ============================================================================
-# MAIN API ENDPOINTS
+# ROUTERS
 # ============================================================================
+#
+# The route handlers live in logos/routers/, grouped by domain. They are
+# imported here, at the bottom of the module, because the routers import the
+# module-level state defined above; importing them at the top would make main
+# and the routers import each other mid-initialisation.
+#
+# The include order matters: user_facing holds the /v1/{path:path} catch-all
+# and is included last, so no route of another router can be shadowed by it.
+# Add new routers above that line.
 
+from logos.routers import admin, internal, logosnode, monitoring, user_facing  # noqa: E402
 
-_AUDIO_BASE_PROPERTIES = {
-    "file": {"type": "string", "format": "binary"},
-    "model": {"type": "string"},
-    "prompt": {"type": "string"},
-    "response_format": {
-        "type": "string",
-        "enum": ["json", "text", "srt", "verbose_json", "vtt"],
-        "default": "json",
-    },
-    "temperature": {"type": "number", "minimum": 0, "maximum": 1},
-}
-
-_TRANSCRIPTION_UPLOAD_REQUEST_SCHEMA = {
-    "required": True,
-    "content": {
-        "multipart/form-data": {
-            "schema": {
-                "type": "object",
-                "required": ["file", "model"],
-                "properties": {
-                    **_AUDIO_BASE_PROPERTIES,
-                    "language": {"type": "string", "description": "ISO-639-1 language code"},
-                    "timestamp_granularities[]": {
-                        "type": "array",
-                        "items": {"type": "string", "enum": ["word", "segment"]},
-                        "description": "whisper-1 with response_format=verbose_json only",
-                    },
-                    "stream": {
-                        "type": "boolean",
-                        "default": False,
-                        "description": "Ignored by whisper-1; supported by newer transcription models",
-                    },
-                },
-            }
-        }
-    },
-}
-
-_TRANSLATION_UPLOAD_REQUEST_SCHEMA = {
-    "required": True,
-    "content": {
-        "multipart/form-data": {
-            "schema": {
-                "type": "object",
-                "required": ["file", "model"],
-                "description": "OpenAI translations currently support the whisper-1 model",
-                "properties": _AUDIO_BASE_PROPERTIES,
-            }
-        }
-    },
-}
-
-_AUDIO_UPLOAD_RESPONSES = {
-    200: {
-        "description": "Transcription result in the requested response format",
-        "content": {
-            "application/json": {},
-            "text/plain": {},
-            "text/vtt": {},
-            "application/x-subrip": {},
-            "text/event-stream": {},
-        },
-    }
-}
-
-
-@app.post(
-    "/v1/audio/transcriptions",
-    tags=["audio"],
-    summary="Create an audio transcription",
-    response_class=Response,
-    responses=_AUDIO_UPLOAD_RESPONSES,
-    openapi_extra={"requestBody": _TRANSCRIPTION_UPLOAD_REQUEST_SCHEMA},
-)
-async def create_audio_transcription(request: Request):
-    """Transcribe an uploaded audio file through an authorized model."""
-    return await handle_sync_request("v1/audio/transcriptions", request)
-
-
-@app.post(
-    "/v1/audio/translations",
-    tags=["audio"],
-    summary="Create an English audio translation",
-    response_class=Response,
-    responses=_AUDIO_UPLOAD_RESPONSES,
-    openapi_extra={"requestBody": _TRANSLATION_UPLOAD_REQUEST_SCHEMA},
-)
-async def create_audio_translation(request: Request):
-    """Transcribe and translate an uploaded audio file into English."""
-    return await handle_sync_request("v1/audio/translations", request)
-
-
-@app.post("/v1/{path:path}", tags=["user-facing"])
-async def logos_service_sync(path: str, request: Request):
-    """
-    Dynamic proxy for OpenAI-compatible API endpoints (/v1/*).
-    Supports both PROXY and RESOURCE modes with streaming.
-
-    POST only: every proxied operation (chat/completions, completions,
-    responses, embeddings, ...) is a POST in the upstream APIs. Other methods
-    get a proper 405 from the router instead of the misleading
-    "400 Invalid JSON body" the body parser used to raise on body-less GETs.
-    """
-    return await handle_sync_request(f"v1/{path}", request)
-
-
-@app.post("/v2/{path:path}", tags=["user-facing"])
-async def logos_service_v2_sync(path: str, request: Request):
-    """
-    Dynamic proxy for Cohere-compatible API endpoints (/v2/embed, /v2/rerank).
-    """
-    return await handle_sync_request(f"v2/{path}", request)
-
-
-@app.post(
-    "/openai/{path:path}",
-    tags=["user-facing"],
-)
-async def logos_service_long_sync(request: Request, path: str = None):
-    """
-    Dynamic proxy for LLM API endpoints (OpenAI-compatible paths).
-    Supports two modes:
-    - PROXY MODE: Direct forwarding to provider (no classification/scheduling)
-    - RESOURCE MODE: Classification + scheduling with SDI-aware pipeline
-
-    :param request: Request object containing headers, body, and client metadata
-    :param path: API endpoint path (e.g., 'chat/completions', 'completions', 'embeddings')
-    :return: StreamingResponse for streaming requests, JSONResponse for synchronous requests
-    """
-    return await handle_sync_request(f"v1/{path}", request)
-
-
-# vLLM non-prefixed endpoints (not part of OpenAI API spec, but user-facing).
-# These are canonical paths for pooling, scoring, reranking, and tokenization.
-async def _handle_vllm_native(request: Request):
-    """Forward to vLLM using the original request path."""
-    path = request.url.path.lstrip("/")
-    return await handle_sync_request(path, request)
-
-
-for _vllm_path in ("/pooling", "/score", "/rerank", "/tokenize", "/detokenize"):
-    app.add_api_route(
-        _vllm_path,
-        _handle_vllm_native,
-        methods=["POST"],
-        tags=["user-facing"],
-        name=f"vllm_native_{_vllm_path.lstrip('/')}",
-    )
-
-
-@app.post(
-    "/jobs/v1/{path:path}",
-    tags=["user-facing"],
-)
-async def logos_service_async(path: str, request: Request):
-    """
-    Async job-based proxy for long running/low-priority requests.
-
-    Params:
-        path: Upstream path to forward.
-        request: Incoming request.
-
-    Returns:
-        202 with job metadata; poll /jobs/{id} for result.
-    """
-    return await submit_job_request(f"v1/{path}", request)
-
-
-@app.post(
-    "/jobs/v2/{path:path}",
-    tags=["user-facing"],
-)
-async def logos_service_v2_async(path: str, request: Request):
-    """Async job-based proxy for Cohere-compatible endpoints."""
-    return await submit_job_request(f"v2/{path}", request)
-
-
-@app.post(
-    "/jobs/openai/{path:path}",
-    tags=["user-facing"],
-)
-async def logos_service_long_async(path: str, request: Request):
-    """
-    Async job-based proxy for OpenAI-compatible, long running/low-priority requests.
-
-    Params:
-        path: Upstream path to forward.
-        request: Incoming request.
-
-    Returns:
-        202 with job metadata; poll /jobs/{id} for result.
-    """
-    return await submit_job_request(f"v1/{path}", request)
-
-
-@app.get("/jobs/{job_id}", tags=["user-facing"])
-async def get_job_status(job_id: int, request: Request):
-    """
-    Return current state of a submitted job, including result or error when finished.
-    Uses team-based authorization - you can only view jobs created by your current team.
-    Logos Admins can view all jobs.
-    """
-    auth = authenticate_api_key(dict(request.headers))
-
-    job = JobService.fetch(job_id)
-    if job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-
-    # Authorization checks
-    job_api_key_id = job.get("api_key_id")
-    job_team_id = job.get("team_id")
-
-    with DBManager() as db:
-        user_info = db.get_user_by_api_key(auth.key_value)
-        is_admin = user_info and user_info.get("role") == "logos_admin"
-
-    if not is_admin:
-        if job_api_key_id != auth.api_key_id:
-            raise HTTPException(status_code=403, detail="Not authorized to access this job")
-        if job_team_id != auth.team_id:
-            raise HTTPException(status_code=403, detail="Job belongs to a different team.")
-
-    return_payload = {
-        "job_id": job_id,
-        "status": job["status"],
-        "result": (job["result_payload"] if job["status"] == JobStatus.SUCCESS.value else None),
-        "error": (job["error_message"] if job["status"] == JobStatus.FAILED.value else None),
-        "created_at": job.get("created_at"),
-        "updated_at": job.get("updated_at"),
-        "team_id": job_team_id,
-    }
-
-    # When a completed job has a non-2xx upstream status code, surface the
-    # error body with the correct HTTP status so OpenAI-spec clients behave
-    # correctly (e.g. don't blind-retry a 400 context-length error).
-    if job["status"] == JobStatus.SUCCESS.value:
-        result_payload = job.get("result_payload") or {}
-        job_status_code = result_payload.get("status_code") if isinstance(result_payload, dict) else None
-        if isinstance(job_status_code, int) and job_status_code >= 400:
-            job_data = result_payload.get("data") or {}
-            corrected_sc, error_body = coerce_upstream_error(job_status_code, job_data)
-            return JSONResponse(
-                content={**return_payload, "result": None, "error": error_body},
-                status_code=corrected_sc,
-            )
-
-    if job["status"] == JobStatus.FAILED.value and job.get("error_message"):
-        # Wrap plain-string failure message in OpenAI error shape
-        _, error_body = coerce_upstream_error(500, {"error": job["error_message"]})
-        return JSONResponse(content={**return_payload, "error": error_body}, status_code=500)
-
-    return return_payload
+app.include_router(monitoring.router)
+app.include_router(internal.router)
+app.include_router(logosnode.router)
+app.include_router(admin.router)
+app.include_router(user_facing.router)
